% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex.navarre}
\alias{ex.navarre}
\title{A spatial polygon with the border of Navarre (Spain)}
\description{
Spatial feature (\code{sf}) representing the border of Navarre with
coordinates in the longitude/latitude format.
}
\keyword{data}
