% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex.ndvi.navarre}
\alias{ex.ndvi.navarre}
\title{An example time-series of NDVI for the region of Navarre in Spain}
\format{The \code{RasterBrick} contains 6 time-periods:
\describe{
  \item{name}{The of each capture contains the date of the period in julian format as YYYYJJJ}
  \item{size}{Each layer contains 113 rows and 105 columns}
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
ex.ndvi.navarre
}
\description{
Geographically projected \code{RasterBrick} object with the NDVI values for the region of Navarre
}
\keyword{datasets}
