% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genFillingIMA.R
\name{genFillingIMA}
\alias{genFillingIMA}
\title{Fills the gaps in a time series of satellite images}
\usage{
genFillingIMA(imgTS, Img2Fill = NULL, aFilter = c(0.05, 0.95),
  fact = 5, nDays = 3, nYears = 1, fun = mean, factSE = 8,
  snow.mode = FALSE, predictSE = FALSE, out.name = "outname", ...)
}
\arguments{
\item{imgTS}{the time series of images to be filled in \code{RasterStack} format 
with the name of the layers containing the date in julian format (\code{YYYYJJJ}).}

\item{Img2Fill}{a vector defining the images to be filled.}

\item{aFilter}{two component vector defining the quantiles to filter the anomaly. Ex. c(0.05,0.95).}

\item{fact}{an aggregation factor of the anomalies before the interpolation.}

\item{nDays}{number of previous and subsequent days used to define the neighborhood}

\item{nYears}{number of years used to define the neighborhood.}

\item{fun}{function used aggregating the anomalies, \code{mean}, \code{median} are acceptable functions.}

\item{factSE}{the fact used in the standard error prediction to reduce the processing time.}

\item{snow.mode}{logical argument. If \code{TRUE} the filling process will be parallelized by \code{raster} package.}

\item{predictSE}{logical argument. If \code{TRUE} the filling process will calculate the standard error instead the prediction.}

\item{out.name}{the name of the output images.}

\item{...}{argument to allow function nestering.
\itemize{
  \item \code{AppRoot} if a path is assigned its save the filled time series of images in this location.
}}
}
\value{
\code{RasterStack} object with the smoothed time series.
}
\description{
\code{genFillingIMA} is the implementation of a spatio-temporal method called Image Mean Anomaly (IMA)
for gap filling published in  \insertCite{militino2019interpolation}{RGISTools}.
}
\details{
The time series of images are decomposed into the mean images and the anomalies images. The procedure applies
the smoothing algorithm over the anomaly images.

The arguments configure the smoothing procedure:
\code{Img2Fill} identifies the images to filled, \code{nPeriods} defines
the spatio temporal neighbourhoood considered when smoothing
a pixel, and \code{fact} sets the level of spatial aggregation.
See more information in \insertCite{militino2019interpolation}{RGISTools}.
}
\examples{
# load an example of NDVI time series in Navarre
data(ex.ndvi.navarre)
# plot data example
genPlotGIS(ex.ndvi.navarre)

# fill the gaps
ndvi.filled<-genFillingIMA(ex.ndvi.navarre,
                           Img2Fill=c(1,2))
# Show the filled images
genPlotGIS(ndvi.filled)
# plot comparison of the cloud and the filled images
ndvi.comp<-stack(ex.ndvi.navarre[[1]],ndvi.filled[[1]],
                 ex.ndvi.navarre[[2]],ndvi.filled[[2]])
genPlotGIS(ndvi.comp,layout=c(2,2))
}
\references{
\insertRef{militino2019interpolation}{RGISTools}
}
