% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genPlotGIS.R
\name{genPlotGIS}
\alias{genPlotGIS}
\title{Plots satellite images with a proper GIS format}
\usage{
genPlotGIS(r, region = NULL, ...)
}
\arguments{
\item{r}{\code{Raster}* class variable with the image or image stack to be plotted.}

\item{region}{\code{polygon} class variable defining the area of interest.}

\item{...}{argument to allow function nestering.
\itemize{
  \item \code{compOpt} list to fit the size and the location of the GIS components as the arrow and the scales.
  \item \code{proj} defines the projection for the plot.
  \item any argument accepted by the \code{spplot} function.
}}
}
\description{
This function displays satellite images with the usual Geographic Information System (GIS) style, 
such as adding the scale, the north arrow, and the borderline of the region of interest.
}
\details{
The function plots any \code{raster} type class as, \code{raster}, \code{RasterStack} or \code{RasterBrick}.
Accepts all parameter used by the function \code{spplot}.
The function \code{genPlotGIS} adds a north arrow and a polygon with the area of interest.
If necessary, the projection of the polygon is transformed to match the projection of the raster.
The projection of the map can be changed by adding the \code{proj} argument. The arrow and scale bar are
located by default in relative positions. Their positions can be modified with the \code{compOpt} paragram.
}
\examples{
# load the ndvi object of Navarre
data(ex.ndvi.navarre)
# load a spatial polygon object of Navarre
data(ex.navarre)

# Shows a panel of 4 maps, one per date. The region of interest is shown in the last map
genPlotGIS(r = ex.ndvi.navarre[[1:4]],
           region = ex.navarre,
           which = c(4), # show region only in the 4th image
           wComp = c(4) # show component only in the 4th image
)

\dontrun{
# Plotting the land surface temperature of the first date available in 2011 in Navarre
# using color palette
library('RColorBrewer')
my.palette <- rev(brewer.pal(n = 9, name = "YlOrRd"))
genPlotGIS(r = ex.ndvi.navarre[[1]],
           region = ex.navarre,
           proj = CRS("+init=epsg:4670"), # project all components
           col.regions = colorRampPalette(my.palette)
)
}

# change scale text relative Y
genPlotGIS(r = ex.ndvi.navarre[[1:4]],
           region = ex.navarre,
           proj = CRS("+init=epsg:4670"),
           compOpt=list(
           # arrow relatives 0-1
           ArrowRelativeX = 0.85,
           ArrowRelativeY = 0.1,
           ArrowRelativeSize = 0.15,
           # scale relatives 0-1
           scaleRelativeX = 0.1,
           scaleRelativeY = 0.1,
           scaleRelativeSize = 0.15,
           #scale text relatives 0-1
           scaleLabelRelativeX = 0.1,
           scaleLabelRelativeY = 0.2,
           scaleLabelSize = 5000
        )
)
}
