% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSaveTSRData.R
\name{genSaveTSRData}
\alias{genSaveTSRData}
\title{Imports time series of images and saves as RData}
\usage{
genSaveTSRData(src, ts.name = "TS.Name", startDate = NULL,
  endDate = NULL, recursive = F, ...)
}
\arguments{
\item{src}{path to the folder where the time series of images is located}

\item{ts.name}{the name of the variable containing the time series in R}

\item{startDate}{the starting date of the time series}

\item{endDate}{the ending date of the time series}

\item{recursive}{a flag to read folders recursively}

\item{...}{argument to allow function nestering
\itemize{
  \item \code{AppRoot} the path where the RData will be saved
}}
}
\description{
\code{genSaveTSRData} imports time series of (tif) satellite images into R from a folder and creates an RData.
}
\details{
The function reads all the images inside the folder specified in \code{src}. The images must be .tif files.
 The \code{src} can take the path created by other functions of this package, such as \code{senMosaic},
 \code{modMosaic}, \code{senFolderToVar}, etc. The images are imported into R to build a \code{RasterStack} that
 is loaded into the global environment. The name of the raster stack will be the one specified
 in ts.name. The RasterStack is saved into an RData file in the Approot directory.
}
\examples{
\dontrun{
# load georeferenced polygon of Navarre as example
data(ex.navarre)
# set the download folder
s.start<-Sys.time()
src<-"Path_for_downloading_folder"
#download the images
modDownload(product="MOD09GA",
            startDate=as.Date("30-07-2018","\%d-\%m-\%Y"),
            endDate=as.Date("06-08-2018","\%d-\%m-\%Y"),
            username="username",
            password="password",
            AppRoot=src,
            hdfdir="hdf",
            tiffdir="tif",
            collection=6,
            extent=ex.navarre)
#set tif folder where hdf will be imported
src<-file.path(src,"MOD09GA")
#set the tif folder path
tif.src<-file.path(src,"tif")
#mosaic and cut navarre region
modMosaic(tif.src,
          AppRoot=src,
          out.name="Navarre",
          extent=ex.navarre)
#change src to navarre folder
src<-file.path(src,"Navarre")
#calculate NDVI from navarre folder
modFolderToVar(src,
               fun=varNDVI,
               AppRoot=dirname(src))
#change src TS_sample
src<-file.path(dirname(src),"NDVI")
#create the Rdata
genSaveTSRData(src,ts.name="ModisNDVI",AppRoot=dirname(src))
s.end<-Sys.time()
}
}
