% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSmoothingCovIMA.R
\name{genSmoothingCovIMA}
\alias{genSmoothingCovIMA}
\title{Uses covariates for smoothing outliers in a time series of satellite images}
\usage{
genSmoothingCovIMA(rStack, cStack, Img2Process = NULL, fun = mean,
  nDays = 3, nYears = 1, fact = 5, out.name = "out",
  aFilter = c(), snow.mode = FALSE, ...)
}
\arguments{
\item{rStack}{a \code{RasterStack} containing a time series of satellite images.}

\item{cStack}{a \code{RasterStack} containing a time series of covariates.}

\item{Img2Process}{a vector defining the images to smooth.}

\item{fun}{a function used to calculate the aggregated image. Both the \code{mean} or \code{median} functions are acceptable.}

\item{nDays}{number of previous and subsequent days used to define the neighborhood}

\item{nYears}{number of years used to define the neighborhood.}

\item{fact}{an aggregation factor to be used to reduce the anomalies before smoothing.}

\item{out.name}{the name of the images if the result is written in the HDD.}

\item{aFilter}{a vector with the lower and upper extreme values for filtering the anomalies. Ex. c(0.05,0.95).}

\item{snow.mode}{logical argument. If \code{TRUE} the filling process will be parallelized by \code{raster} package.}

\item{...}{argument to allow function nestering.
\itemize{
  \item \code{AppRoot} the path where the images will be saved in tif format.
}}
}
\description{
\code{genSmoothingCovIMA} runs the IMA (\link{genSmoothingIMA}) smoothing method with covariates 
\insertCite{militino2019interpolation}{RGISTools}.
}
\details{
The time series of images are decomposed into the mean images and the anomaly images. The procedure applies
the smoothing algorithm over the anomaly images. The procedure smoothed the anomaly using covariates to
smooth the outliers in the anomaly. \insertCite{militino2019interpolation}{RGISTools}.

This function provides a set of arguments to configure the run of the precedure, such as the images to fill (\code{Img2Process}),
the number of periods or years in the definition of the spatio temporal neighbourhoood, or the size of the
aggregation. All these particularities are explained in the publication.
}
\examples{
set.seed(0)
# load example ndvi and dem data of Navarre
data(ex.ndvi.navarre)
data(ex.dem.navarre)
# plot example data
genPlotGIS(ex.ndvi.navarre)
genPlotGIS(ex.dem.navarre)

# distorts 5\% of the original ndvi data by 
# altering 50\% its values
for(x in c(2,5)){
  aux <- sampleRandom(ex.ndvi.navarre[[x]],
                      ncell(ex.ndvi.navarre) * 0.05,
                      cells = TRUE,
                      na.rm = TRUE)
  ex.ndvi.navarre[[x]][aux[,1]] <- aux[,2] * 1.5
}
genPlotGIS(ex.ndvi.navarre)

# smoothing the image using dem as covariate
smth.ndvi <- genSmoothingCovIMA(rStack = ex.ndvi.navarre,
                                cStack = ex.dem.navarre,
                                Img2Process = c(2,5))
# plot the distorted 1, smoothed 1, 
# distorted 5, smoothed 5 images
plot(stack(ex.ndvi.navarre[[2]],
                 smth.ndvi[[1]],
                 ex.ndvi.navarre[[5]],
                 smth.ndvi[[2]]))
}
\references{
\insertRef{militino2018improving}{RGISTools}
}
