% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSmoothingIMA.R
\name{genSmoothingIMA}
\alias{genSmoothingIMA}
\title{Fill the gaps in a time series of satellite images}
\usage{
genSmoothingIMA(imgTS, Img2Fill = NULL, aFilter = c(0.05, 0.95),
  fact = 5, nDays = 3, nYears = 1, fun = mean, factSE = 8,
  snow.mode = FALSE, predictSE = FALSE, out.name = "outname", ...)
}
\arguments{
\item{imgTS}{\code{RasterStack} class argument containing the time series of images to be filled
with the name of the layers containing the date in julian format (\code{YYYYJJJ}).}

\item{Img2Fill}{\code{vector} class argument defining the images to be filled.}

\item{aFilter}{\code{vector} class argument with two component defining the quantiles to filter the anomaly. Ex. c(0.05,0.95).}

\item{fact}{\code{numeric} argument with an aggregation factor of the anomalies before the interpolation.}

\item{nDays}{\code{numeric} argument with the number of previous and subsequent days used for defining the neighborhood}

\item{nYears}{\code{numeric} argument with the number of previous and subsequent years used for defining the neighborhood.}

\item{fun}{function used aggregating the anomalies, \code{mean}, \code{median} are acceptable functions.}

\item{factSE}{the fact used in the standard error prediction to reduce the processing time.}

\item{snow.mode}{logical argument. If \code{TRUE}, the filling process will be parallelized by \code{raster} package.}

\item{out.name}{the name of the output images.}

\item{...}{argument for function nestering.
\itemize{
  \item \code{AppRoot} the path where the filled time series of images will be saved in \code{GTiff} format.
}}
}
\value{
\code{RasterStack} object with the smoothed time series.
}
\description{
\code{genSmoothingIMA} is the implementation of a spatio temporal method called Image Mean Anomaly (IMA)
for gap filling \insertCite{militino2019interpolation}{RGISTools}.
}
\details{
The time series of images are decomposed into time series of the mean images and time series of the anomaly images. The procedure applies
the smoothing algorithm over the anomaly images.

The arguments configure the smoothing procedure:
\code{Img2Fill} identifies the images to fill, \code{nDays} and \code{nYears} defines
the spatio temporal neighbourhoood of the target images considered, and \code{fact} sets the 
level of spatial aggregation \insertCite{militino2019interpolation}{RGISTools}.
}
\examples{
# load an example of NDVI time series in Navarre
data(ex.ndvi.navarre)
# the 2 images to be filled and the neighbourhood
genPlotGIS(ex.ndvi.navarre)

# Filled images
ndvi.filled <- genSmoothingIMA(ex.ndvi.navarre,
                               Img2Fill = c(1,2))
# Show the filled images
genPlotGIS(ndvi.filled)
# plot comparison of the cloud and the filled images
ndvi.comp <- stack(ex.ndvi.navarre[[1]], ndvi.filled[[1]],
                   ex.ndvi.navarre[[2]], ndvi.filled[[2]])
genPlotGIS(ndvi.comp, layout=c(2, 2))
}
\references{
\insertRef{militino2019interpolation}{RGISTools}
}
