% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls7FolderToVar.R
\name{ls7FolderToVar}
\alias{ls7FolderToVar}
\title{Computes derived variables from Landsat-7 multispectral bands}
\usage{
ls7FolderToVar(src, fun, getStack = FALSE, overwrite = FALSE, ...)
}
\arguments{
\item{src}{path to the folder with the Landsat multispectral image.}

\item{fun}{is a function defined by the package for computing indexes.
All functions created in RGISTools starting with
'var' are avaliable functions. Custom functions can be also
implemented \code{var} are acceptable functions.}

\item{getStack}{logical argument. If \code{TRUE}, returns the time-series as a 
\code{RasterStack}, otherwise as Hard Drive Devide (HDD).}

\item{overwrite}{logical argument. If \code{TRUE} overwrites the existing images with the same name.}

\item{...}{argument to allow function nestering.
\itemize{
  \item \code{AppRoot} the directory of the resulting time series.
}}
}
\description{
\code{\link{ls7FolderToVar}} computes indexes using the bands from Landsat-7 multispectral
images. The images are specified by a path to the storing folder (resulting from
the \code{\link{lsMosaic}} function). The function returns a \code{RasterStack} with the time-series of the indexes.
}
\details{
The function requires to define \code{src} and \code{fun} attributes. \code{src} defines the path to
the result of \code{\link{lsMosaic}}, with all bands of Landsat-7 for a region of interest. \code{fun} defines
the variable of interest using any of the functions in the packages starting with \code{var} (\code{\link{varNDVI}},
 \code{\link{varEVI}}, ...)
}
\examples{
\dontrun{
# load a spatial polygon object of navarre for the example
data(ex.navarre)
# asign the folder where the example will be run
src <- "Path_for_downloading_folder"
# download Landsat-7 images
lsDownload(satellite = "ls7",
           username = "username",
           password = "password",
           startDate = as.Date("01-01-2018", "\%d-\%m-\%Y"),
           endDate = as.Date("20-01-2018", "\%d-\%m-\%Y"),
           extent = ex.navarre,
           untar = TRUE,
           AppRoot = src)
# asign the folder with the Landsat-7 images untared
tif.src <- file.path(src, "untar")
# mosaic the Landsat7 images
lsMosaic(tif.src,
         AppRoot = src,
         out.name = "Navarre")
# asign src as the path to mosaicked folder
src2 <- file.path(src, "Navarre")
# generate NDVI images of Navarre
src3 <- file.path(src1, "Navarre_Variables")
dir.create(src3)
ls7FolderToVar(src2,
               fun = varNDVI,
               AppRoot = src3,
               overwrite = T)
               
flist <- list.files(file.path(src3,"EVI"),
                    pattern = "\\\\.tif$",
                    full.names = TRUE,
                    recursive = TRUE)

files.raster <- stack(flist)
spplot(files)
}
}
