% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls7LoadMetadata.R
\name{ls7LoadMetadata}
\alias{ls7LoadMetadata}
\title{Sinchronizes landsat 7 meta data file for image search.}
\usage{
ls7LoadMetadata(update = F, verbose = T, omit.question = F, ...)
}
\arguments{
\item{update}{TRUE/FALSE argument to force update of metadata file}

\item{verbose}{TRUE/FALSE argument to print all steps of metadata download process}

\item{omit.question}{omit the question for ensure the loading of the metadata}

\item{...}{accepts \code{AppRoot} as root directory where meta data file will be saved
or/and other argument for function nestering}
}
\description{
\code{ls7LoadMetadata} loads a data frame called \code{.LS7MD}
 with the names of the Landsat-7 images and their metadata.The metadata provides
 auxiliary information regarding Landsat-7 images repository such as image quality, acquisition
 data, cloud cover, etc. You can find a description of the metadata at
 \url{https://www.usgs.gov/land-resources/nli/landsat/bulk-metadata-service}
}
\details{
All captures done by Landsat-7 are cataloged in a unique csv file. The size of the file
might be larger than 360MB. Therefore, the process of downloading and importing R may take several
minutes (around 15 minutes in a Intel Core i7-4790, 16Gb of RAM and Hard Drive Device). The function creates an
RData file with the metadata csv. Thus, every time \code{ls7MetaDataFile} is called,
this function loads the existing RData in the Approot. This is intended to reduce the loading
time of metadata in the future.
}
\examples{
\dontrun{
#creates a MetaData folder and downloads the csv on working directory
ls7LoadMetadata()

#creates a MetaData folder and downloads the csv on "C:/LandsatDownload" directory
ls7LoadMetadata(AppRoot="C:/LandsatDownload")

Force renew existing meta data csv
ls7LoadMetadata(update=TRUE)
}
}
