% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls7Search.R
\name{ls7Search}
\alias{ls7Search}
\title{Search Landsat-7 time series list of images}
\usage{
ls7Search(startDate, endDate, verbose = FALSE, precise = FALSE, ...)
}
\arguments{
\item{startDate}{starting date of the image time series in \code{Date} class. For instance, using any format from \code{as.Date} function.}

\item{endDate}{ending date of the image time series in \code{Date} class. For instance, using any format from \code{as.Date} function.}

\item{verbose}{logical argument. If \code{TRUE}, the function prints running stages and warnings.}

\item{precise}{logical argument. If \code{TRUE}, the search is done tile by tile (slower).}

\item{...}{argument for function nestering:
 \itemize{
  \item \code{pathrow} a list of vectors defining the path and row number for the region of interest according
to the \href{https://landsat.gsfc.nasa.gov/the-worldwide-reference-system/}{Worldwide Reference System}.
This argument is mandatory if \code{extent} is not defined. Ex. \code{c(-1.64323,42.81687)}.
  \item \code{lonlat} this argument is optional. A vector with the coordinates of
the point or region of interest in longitude/latitude format.
  \item \code{extent} this argument is optional. Extent, Raster*, SpatialPolygons*, SpatialLines* or SpatialPoints*
object are acceptable formats as long as are longitude/latitude format.
This argument is mandatory if \code{pathrow} is not defined.
  \item \code{AppRoot} the root directory where metadata file will be saved.
  \item all column names in .LS7MD data frame for filter results.
}}
}
\description{
\code{ls7Search} searches the Landsat-7 image repository to find available images for
a particular location and date interval. The function returns the search result as a data frame
with the names of the images and their metadata.
}
\details{
\code{ls7Search} is a stand-alone function. If the metadata for the time and region of interest has been
downloaded before, \code{ls7Search} will use this metadata by default. In case the metadata has not
been yet downloaded, \code{ls7Search} will make the call for you.

The search is done by defining a time interval and a location. The arguments \code{startDate}
and \code{endDate} defines the time interval. These are mandatory arguments. The function defines the spatial location
using at least one of the following arguments: \code{pathrow}, \code{extent}, \code{lonlat}, and \code{polygon}. 
When more than one of these argument are defined,
the function will work with the first evaluated method, otherwise, the function shows an error message.

Landsat images are catalogued spatially using a unique path and row. The fastest way to search an image
in the metadata file is filtering by its path and row. This search method requires previous knowledge on
the path and row relevant for your region of interest.

From the user point of view, the easiest way to search a time series of Landsat-7 is using the extent,
\code{lonlat} and \code{polygon} arguments. These methods do not require to know in advance the path and rows of the images.
These method uses spatial objects to define the region of interest, and needs to be
"\code{+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs}”. The argument \code{extent} accepts any R object being defined
by a spatial extent. The argument \code{lonlat} only accepts an R vector with a point in the form of
longitude/latitude (ex. \code{c(-1.64323,42.81687)}).
The argument \code{Polygon}, accepts \code{SpatialPolygon} or \code{SpatialPolygonDataFrame} objects.

In addition, the search function enables further filtering. The function can filter
the results by any column name in the metadata file, using the column name as an argument. For example, to
filter the images that can be previewed, the user has to find the images with “Y” in the browseAvaliable column.
This can be achieved by adding \code{browseAvaliable=”Y”} as a function argument.
}
\examples{
\dontrun{
# search by known row and path
search.res <- ls7Search(startDate = as.Date("01-01-2011", "\%d-\%m-\%Y"),
                        endDate = as.Date("31-12-2013", "\%d-\%m-\%Y"),
                        pathrow = list(c(200,31),c(200,30)),
                        browseAvaliable = "Y")
                  
# search by longitude/latitude
search.res <- ls7Search(startDate = as.Date("01-01-2011", "\%d-\%m-\%Y"),
                        endDate = as.Date("31-12-2013", "\%d-\%m-\%Y"),
                        lonlat = c(-1.64323,42.81687),
                        browseAvaliable = "Y")
                  
# search by projected file must be in long/lat projection
# load a spatial polygon object of Navarre
data(ex.navarre)
search.res <- ls7Search(startDate = as.Date("01-01-2011", "\%d-\%m-\%Y"),
                        endDate = as.Date("31-12-2013", "\%d-\%m-\%Y"),
                        extent = ex.navarre,
                        precise = TRUE,
                        browseAvaliable = "Y")

# search by projected file fast
search.res <- ls7Search(startDate = as.Date("01-01-2011","\%d-\%m-\%Y"),
                        endDate = as.Date("31-12-2013","\%d-\%m-\%Y"),
                        extent = ex.navarre,
                        precise = FALSE,
                        browseAvaliable = "Y")
# remove metadata data frame to free memory
lsRemoveMetadata()
}
}
