% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls8FolderToVar.R
\name{ls8FolderToVar}
\alias{ls8FolderToVar}
\title{Compute derived variables from Landsat-8 multispectral bands}
\usage{
ls8FolderToVar(src, fun, getStack = FALSE, overwrite = FALSE, ...)
}
\arguments{
\item{src}{path to the folder with the Landsat multispectral image.}

\item{fun}{is a function defined for computing indexes.
All functions created in RGISTools starting with
'var' are avaliable functions. Custom functions can be also
implemented \code{var} are acceptable functions.}

\item{getStack}{logical argument. If \code{TRUE}, returns the 
time-series of images as a \code{RasterStack}, otherwise the images are saved in the Hard Drive Devide (HDD).}

\item{overwrite}{logical argument. If \code{TRUE}, overwrites the 
existing images with the same name.}

\item{...}{argument for function nestering
\itemize{
  \item \code{AppRoot} the directory of the resulting time series
}}
}
\description{
\code{ls8FolderToVar} calculates an index using the bands from Landsat-8 multispectral
images. The images are specified by a path to the storing folder (resulting from
the \code{\link{lsMosaic}} function). The function returns a \code{RasterStack} with the index time series.
}
\details{
The function requires to define \code{src} and \code{fun} input attributes. \code{src} defines the path to
the result of \code{\link{lsMosaic}}, with all bands of Landsat-8 for a region of interest. \code{fun} defines
the variable of interest using any of the functions in the packages starting with \code{var} prefix (\code{\link{varNDVI}},
 \code{\link{varEVI}}, ...)
}
\examples{
\dontrun{
# load a spatial polygon object of Navarre
data(ex.navarre)
# assign the main output directory
src <- "Path_for_downloading_folder"
# download Landsat-8 images
lsDownload(satellite = "ls8",
           username = "username",
           password = "password",
           startDate = as.Date("01-01-2018","\%d-\%m-\%Y"),
           endDate = as.Date("20-01-2018","\%d-\%m-\%Y"),
           extent = ex.navarre,
           untar = TRUE,
           AppRoot = src)
# assign the folder with the Landsat-8 untared images
tif.src <- file.path(src, "untar")
# mosaic the Landsat-8 images
lsMosaic(tif.src,
         AppRoot = src,
         out.name = "Navarre")
# assign src as the path to mosaiced folder
src2 <- file.path(src, "Navarre")
# generate NDVI images of Navarre
src3 <- file.path(src1, "Navarre_Variables")
dir.create(src3)
ls8FolderToVar(src2,
               fun = varNDVI,
               AppRoot = src3,
               overwrite = T)
               
flist <- list.files(file.path(src3,"EVI"),
                    pattern = "\\\\.tif$",
                    full.names = TRUE,
                    recursive = TRUE)

files.raster <- stack(flist)
spplot(files)
}
}
