% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls8LoadMetadata.R
\name{ls8LoadMetadata}
\alias{ls8LoadMetadata}
\title{Sinchronizes Landsat-8 meta data file, for image search}
\usage{
ls8LoadMetadata(update = FALSE, verbose = TRUE,
  omit.question = FALSE, ...)
}
\arguments{
\item{update}{logical argument. If \code{TRUE} is mandatory to update the metadata file.}

\item{verbose}{logical argument. If \code{TRUE} the function prints running stages and warnings.}

\item{omit.question}{logical argument. If \code{TRUE} omits the question for ensuring the loading of the metadata file.}

\item{...}{argument to allow function nestering
\itemize{
  \item \code{AppRoot} the directory where the metadata will be saved.
}}
}
\description{
\code{ls8LoadMetadata} downloads the csv format file of Metadata and loads a hidden data frame called .LS8MD into R.The metadata provides
 auxiliary information regarding Landsat-8 images repository such as image quality, acquisition
 data, cloud cover, etc. You can find a description of the metadata at
 \href{https://www.usgs.gov/land-resources/nli/landsat/bulk-metadata-service}{bulk-metadata-service}.
}
\details{
All captures done by Landsat-8 are cataloged in a unique csv file. The size of the file
might be larger than 210MB. Therefore, the process of downloading and importing to R may take several
minutes (around 7 minutes in a Intel Core i7-4790, 16Gb of RAM and Hard Drive storage). The function creates an
RData file with the metadata csv. Thus, every time \code{ls8LoadMetadata} is called,
this function loads the existing RData in the \code{AppRoot}.
}
\examples{
\dontrun{
# creates a MetaData folder and downloads the csv on working directory
ls8LoadMetadata()

# creates a MetaData folder and downloads the csv on "Path_for_metadata" directory
ls8LoadMetadata(AppRoot = "Path_for_metadata")

# Force renew existing meta data csv
ls8LoadMetadata(update = TRUE)
}
}
