% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls8Search.R
\name{ls8Search}
\alias{ls8Search}
\title{Search Landsat-8 time-series images list}
\usage{
ls8Search(startDate, endDate, verbose = FALSE, precise = FALSE, ...)
}
\arguments{
\item{startDate}{starting date of the image time series in \code{Date} class. For instance, using any format from \code{as.Date} function.}

\item{endDate}{ending date of the image time series in \code{Date} class. For instance, using any format from \code{as.Date} function.}

\item{verbose}{logical argument. If \code{TRUE} the function prints running stages and warnings.}

\item{precise}{logical argument. If \code{TRUE} the search is donw tile by tile (slower).}

\item{...}{argument to allow function nestering:
 \itemize{
  \item \code{pathrow} a list of vectors defining the path and row number for the region of interest according
to the \href{https://landsat.gsfc.nasa.gov/the-worldwide-reference-system/}{Worldwide Reference System}
This argument is mandatory if extent is not defined.
  \item \code{lonlat} this argument is optional. A vector or a polygon with the coordinates of
the point or region of interest in longitude/latitude format.
  \item \code{extent} this argument is optional. \code{Extent}, \code{Raster*}, \code{SpatialPolygons*}, \code{SpatialLines*} or \code{SpatialPoints*}
object are acceptable formats as long as are longitude/latitude format.
This argument is mandatory if pathrow is not defined.
  \item \code{AppRoot} the root directory where meta data file will be saved.
  \item all column names in .LS8MD data frame for filter results.
}}
}
\description{
\code{ls8Search} searches the Landsat-8 image repository to find those which are relevant for
a particular location and date interval. The function returns the search result as a data frame
with the names of the images and their metadata.
}
\details{
\code{ls8Search} is a stand-alone function. If the metadata for the time and region of interest has been
downloaded before, \code{ls8Search} will use this metadata by default. In case the metadata has not
been yet downloaded, \code{ls8Search} will make the call for you.

The search is done by defining a temporal interval and a location. The arguments \code{startDate}
and \code{endDate} defines the temporal interval. These are mandatory arguments. The function defines the spatial location
using at least one of the following arguments: \code{pathrow}, \code{extent}, \code{lonlat} and \code{polygon}. 
When more than one of these argument are defined,
the function will work with the first evaluated method, otherwise, the function shows an error message.

Landsat images are catalogued spatially using a unique path and row. The fastest way to search an image
in the metadata file is filtering by its path and row. This search method requires previous knowledge on
the path and row relevant for your region of interest.

From the user point of view, the easiest way to search a time series of Landsat-8 is using the extent,
\code{lonlat} and polygon arguments. These methods do not requires to know in advance the path and rows of the images.
These method uses spatial objects to define the region of interest, and needs to be
"\code{+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs}”. The argument \code{extent} accepts any R object being defined
by a spatial extent. The argument \code{lonlat} only accepts an R vector with one coordinate in the form of
longitude/latitude (ex. \code{c(-1.64323,42.81687)}).
The argument \code{Polygon}, accepts \code{SpatialPolygon} or \code{SpatialPolygonDataFrame} objects.

In addition, the search function enables further filtering. The function can filter
the results by any column name in the metadata file, using the column name as an argument. For example, to
filter the images that can be previewed, the user has to find the images with “Y” in the browseAvaliable column.
This can be achieved by adding \code{browseAvaliable=”Y”} as a function argument.
}
\examples{
\dontrun{
# search by known row and path
search.res <- ls8Search(startDate = as.Date("01-01-2011", "\%d-\%m-\%Y"),
                        endDate = as.Date("31-12-2013", "\%d-\%m-\%Y"),
                        pathrow = list(c(200, 31), c(200, 30)),
                        browseAvaliable = "Y")

# search by projected file must be in lat long projection
data(ex.navarre)
search.res <- ls8Search(startDate = as.Date("01-01-2011", "\%d-\%m-\%Y"),
                        endDate = as.Date("31-12-2013", "\%d-\%m-\%Y"),
                        extent = ex.navarre,
                        precise = TRUE,
                        browseAvaliable = "Y")

# search by projected file fast
search.res <- ls8Search(startDate = as.Date("01-01-2011", "\%d-\%m-\%Y"),
                      endDate = as.Date("31-12-2013", "\%d-\%m-\%Y"),
                      extent = ex.navarre,
                      precise = FALSE,
                      browseAvaliable = "Y")
# remove metadata data frame to free memory
lsRemoveMetadata()
}
}
