% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsDownload.R
\name{lsDownload}
\alias{lsDownload}
\title{Search and downloads a time series of satellite images from Landsat 7-8}
\usage{
lsDownload(satellite, startDate, endDate, username, password,
  verbose = FALSE, untar = TRUE, raw.rm = FALSE, ...)
}
\arguments{
\item{satellite}{string containing the type of satellite for downloading (\code{"ls7"} or \code{"ls8"}).}

\item{startDate}{starting date of the image time series in \code{Date} class. For instance, using any format from \code{as.Date} function.}

\item{endDate}{ending date of the image time series in \code{Date} class. For instance, using any format from \code{as.Date} function.}

\item{username}{login credentials to access the USGS EROS web service.}

\item{password}{login credentials to access the USGS EROS web service.}

\item{verbose}{logical argument. If \code{TRUE}, the function prints running stages and warnings.}

\item{untar}{logical argument. If \code{TRUE}, untars downloaded images.}

\item{raw.rm}{logical argument. If \code{TRUE}, removes the raw images.}

\item{...}{argument for function nestering accepts:
 \itemize{
  \item any argument for function nestering including all \code{\link{ls8Search}/\link{ls7Search}} or \code{\link{lsDownSearch}}
function arguments.
  \item \code{AppRoot} the directory to save the downloaded images.
}}
}
\description{
\code{lsDownSearch} downloads the list of images provided by \code{ls7Search} and \code{ls8Search} functions.
The images are saved as ‘.tiff’ files in the \code{AppRoot} directory.
}
\details{
This function is used for searching and downloading Landsat images. The function uses landsat
search function (\code{\link{ls7Search}} or \code{\link{ls8Search}}) depending on the \code{satellite} argument,
and then, calls \code{\link{lsDownload}} function with the response of the search.
Image download requires USGS login account. \href{https://ers.cr.usgs.gov/register/}{Get your credentials}.

The image files from the USGS EROS web service are compressed as ‘.tar.gz’ files. \code{\link{lsDownload}} decompresses the
images and obtains the corresponding ‘.tif’ files. The ‘.tif’ files are saved in the
\code{AppRoot} directory. When \code{untarDir} is defined the function untars the images in this folder.
This replicates the images in compresses version as tar.gz and uncompresses version. To save space in the disk
\code{raw.rm = T} can be defined, and \code{\link{lsDownload}} will remove the ‘tar.gz’ files.
If \code{raw.rm = F}, the original files remain, which might be useful to have access to the original files
in the future and avoid further downloads. By default, \code{\link{lsDownload}} saves the images in (...), in the \code{AppRoot}
directory. To change this setting, provide \code{AppRoot = "the full path as an argument"}.
}
\examples{
\dontrun{
# load a spatial polygon object of Navarre
data(ex.navarre)

# Search and download the images from Landsat-8 comprised between
# 01-01-2018 and 20-01-2018 in the region of Navarre
lsDownload(satellite = "ls8",
           username = "username",
           password = "password",
           startDate = as.Date("01-01-2018", "\%d-\%m-\%Y"),
           endDate = as.Date("20-01-2018", "\%d-\%m-\%Y"),
           extent = ex.navarre)
           
# remove metadata data frame to free memory
lsRemoveMetadata()
}
}
