% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsEspaUpdateOrders.R
\name{lsEspaUpdateOrders}
\alias{lsEspaUpdateOrders}
\title{Title}
\usage{
lsEspaUpdateOrders(images, username = NULL, password = NULL,
  c.handle = NULL)
}
\arguments{
\item{images}{}

\item{username}{}

\item{password}{}

\item{c.handle}{}
}
\value{

}
\description{
Title
}
