% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsGetPathRow.R
\name{lsGetPathRow}
\alias{lsGetPathRow}
\title{Return the pathrow of a tile of Landsat-7 or Landsat-8 images}
\usage{
lsGetPathRow(str)
}
\arguments{
\item{str}{path to the folder with the Landsat-7 or 
Landsat-8 multispectral images.}
}
\value{
a string with the path and row in "\code{ppprrr}" format.
}
\description{
\code{lsGetPathRow} reads the official name of a Landsat-7 or Landsat-8 image
and returns the path and row of its tile.
}
\examples{
# example of getting date from Landsat-8 image name
str <- c("LE72000302011066ASN00",
         "LE72000302011066ASN00")
pr <- lsGetPathRow(str)
print(pr)

}
