% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsRemoveMetadata.R
\name{lsRemoveMetadata}
\alias{lsRemoveMetadata}
\title{Removes from R environment LS7 and LS8 metadata files}
\usage{
lsRemoveMetadata()
}
\description{
\code{lsRemoveMetadata} removes metadata from R enviroment.
}
\details{
If the metadata file has been loaded by any Landsat search function or using load
metadata function, \code{lsRemoveMetadata} removes both dataframes \code{.LS8MD} and/or \code{.LS7MD}.
}
\examples{
\dontrun{
# creates a MetaData folder and downloads the csv on working directory
ls8LoadMetadata()
lsRemoveMetadata()
}
}
