% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genMissionGetDates.R
\name{modGetDates}
\alias{modGetDates}
\title{Returns the capturing date of Modis image}
\usage{
modGetDates(str, ...)
}
\arguments{
\item{str}{\code{character} class or a \code{vector} class containing the file path(s) or file name(s).}

\item{...}{argument to allow function nestering:
\itemize{
  \item \code{date.format} modify the format of the date being returned.
}}
}
\description{
\code{modGetDates} extracts the capturing date of one or several Modis
images, given their file paths or names. The function returns a date class.
}
\details{
The function works with names or file paths having either of the two file extensions,
 \code{HDF} or \code{GTiff} files. The function accepts more than on file path or file names.
 If so, they should be provided as a list. Dates are returned as '\code{YYYY-mm-dd}' by default.
 If another format is required, it can be modified through the argument \code{date.format}.
}
\examples{
# example of getting date from Sentinel2 image name
imgPath <- 'MYD13A2.A2016361.h17v04.006.2017285133407.hdf'
modGetDates(imgPath)

# example of a list of the full file paths of Modis images, mixing .hdf and .tif files
imgsPaths<-list('MYD13A2.A2013297.h17v04.006.2015269230726.hdf',
                'MYD13A2.A2013313.h17v04.006.2015271071143.tif')
modGetDates(imgsPaths, date.format = "\%Y\%j")
}
