% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{navarre}
\alias{navarre}
\title{Geographically projected spatial polygon of the region of Navarre in Spain}
\format{An object of class \code{SpatialPolygonsDataFrame} with 1 rows and 3 columns.}
\usage{
navarre
}
\description{
\code{SpatialPolygonsDataFrame} containing the region of Navarre projected in latitude/longitude
}
\keyword{datasets}
