% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senDownSearch.R
\name{senDownSearch}
\alias{senDownSearch}
\title{Downloads Sentinel images from search function response}
\usage{
senDownSearch(searchres, username, password,
  error.log = "download_error.log", nattempts = NULL, unzip = F,
  overwrite = F, ...)
}
\arguments{
\item{searchres}{response from \code{senSearch}}

\item{username}{login credentials to access the ESA’s SciHub web service}

\item{password}{login credentials to access the ESA’s SciHub web service}

\item{error.log}{error log file name}

\item{nattempts}{the number of attempts that the function has to carry out to download an image in case the file becomes corrupted.}

\item{unzip}{flag for unzipping the images}

\item{overwrite}{flag for overwriting the resulting the images}

\item{...}{argument to allow function nestering
\itemize{
  \item \code{AppRoot} the directory to save the resulting time series
}}
}
\description{
\code{senDownSearch} downloads the list of urls generated y the function senSearch by using the ESA’s SciHub API
}
\details{
\code{senDownSearch} downloads the images from Sentinel products using the search results provided by \code{\link{senSearch}}.
 The raw images are downloaded into the Aproot directory. In case the download is interrupted,
 the image file could be corrupted. The function detects the corrupted files to restart the process.
 To prevent the computer from crashing, the nattempts  flag limits the number of attempts to download the image.
 The default number of attempts is set to 3.
senDownload requires the credentials to access the ESA’s SciHub data service. Please,
sign up at: \url{https://scihub.copernicus.eu/dhus/#/self-registration}
}
\examples{
\dontrun{
# Download S2MSI1C products sensed by Sentinel - 2 satellite in July-August 2018
data(ex.navarre)
searchres<-senSearch(startDate=as.Date("2018-07-29","\%Y-\%m-\%d"),
                     endDate=as.Date("2018-08-06","\%Y-\%m-\%d"),
                     platform="Sentinel-2",
                     extent=ex.navarre,
                     product="S2MSI1C",
                     username="username",
                     password="password")

#filtering the path R094 where Navarre is located
length(searchres)
searchres<-searchres[grepl("R094",names(searchres))]
length(searchres)

#sentinel download function
senDownSearch(searchres=searchres,
              username="username",
              password="password",
              AppRoot="Path_for_downloading_folder",
              unzip=T)
}
}
