% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senGetPathRow.R
\name{senGetTile}
\alias{senGetTile}
\title{Returns the tile of Sentinel-2 images "TTTSSS" string format}
\usage{
senGetTile(str)
}
\arguments{
\item{str}{the full path or official image name of the Sentinel image from which the tile is extracted.}
}
\value{
an string with the path and row of the image in "\code{TTTSSS}" format.
}
\description{
\code{senGetTile} reads the official name of a Sentinel image and returns its assigned tile. 
\href{https://sentinel.esa.int/web/sentinel/missions/sentinel-2/news/-/asset_publisher/Ac0d/content/sentinel-2-level-1c-product-tiling-grid-released}{Get more Sentinel tile detalis here}.
}
\examples{
#example of getting date from Sentinel2 image name
str <- c("S2A_MSIL1C_20170102T111442_N0204_R137_T30TWN_20170102T111441.SAFE",
         "S2A_OPER_PRD_MSIL1C_PDMC_20160308T090616_R094_V20160305T110109_20160305T110109")
pr <- senGetTile(str)
print(pr)

}
