% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senSearch.R
\name{senSearch}
\alias{senSearch}
\title{Search Sentinel images using ESA’s SciHub API}
\usage{
senSearch(username, password, ...)
}
\arguments{
\item{username}{Scihub username}

\item{password}{Scihub password}

\item{...}{argument to allow function nestering
\itemize{
  \item \code{product} Sentinel product type
  \item \code{startDate} Starting date of time series
  \item \code{endDate} Ending date of time series
  \item \code{extent} Location as projecte file with extention
  \item \code{platform} platform name of the Sentinel mission (Sentine-1, Sentinel-2,...)
  \item \code{qformat} format of response
  \item \code{verbose} Debugging flag
  \item \code{error.log} error log file name
  \item \code{AppRoot} the directory where the images will be saved
}}
}
\description{
\code{senSearch} senSearch searches sentinel products on ESA powered api  \url{http://scihub.copernicus.eu}.
}
\details{
Provides the images and urls to be downloaded from ESA’s SciHub API. The images are
searched based on a range of dates and a region of interest. Dates have to be
provided as a date class object. The region of interest should be provided as a
spatial class object with an EPSG:4326 coordinate projection. ESA’s API only returns
as much as 100 entries.
Credentials from ESA’s SciHub are needed to use this function.
Please, sign up at: \url{https://scihub.copernicus.eu/dhus/#/self-registration}

For further information on ESA’s missions and data products, please visit:
\url{https://sentinel.esa.int/web/sentinel/missions}
}
\examples{
\dontrun{
#load a spatial polygon object of navarre for the example
data(ex.navarre)
#perform the search query
searchres<-senSearch(startDate=as.Date("2018210","\%Y\%j"),
                     endDate=as.Date("2018218","\%Y\%j"),
                     platform="Sentinel-2",
                     extent=ex.navarre,
                     product="S2MSI1C",
                     username="username",
                     password="password")
head(searchres)
}
}
