% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGISToolsOpts.R
\name{setRGISToolsOpt}
\alias{setRGISToolsOpt}
\title{Changes the default value of an RGISTools option}
\usage{
setRGISToolsOpt(opt, value, env = optEnv)
}
\arguments{
\item{opt}{the name of the option to change in character format}

\item{value}{the new value of selected option}

\item{env}{the environment where the RGISTools option are saved}
}
\description{
\code{setRGISToolsOpt} changes the default value of an RGISTools option.
}
\examples{
getRGISToolsOpt("LS7META.dir")
setRGISToolsOpt("LS7META.dir","NewMTDir")
getRGISToolsOpt("LS7META.dir")
}
