% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex.dem.navarre}
\alias{ex.dem.navarre}
\title{Time series of Digital Elevation Model (DEM) for the region of Navarre in Spain}
\format{The \code{RasterStack} contains 6 time-periods, with the same DEM:
\describe{
  \item{name}{each layer is named with the date of the period in julian format as "\code{YYYYJJJ}"}.
  \item{size}{113 rows by 105 columns and 6 layers}.
}}
\description{
Geographically projected \code{RasterStack} with the DEM values for
the region of Navarre, to use as example in \code{genSmoothingIMA}.
}
\keyword{data}
