% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex.dem.navarre}
\alias{ex.dem.navarre}
\title{An example time-series of Digital Elevation Model (DEM) for the region of Navarre in Spain}
\format{The \code{RasterBrick} contains 6 time-periods, with the same DEM:
\describe{
  \item{name}{The of each capture contains the date of the period in julian format as YYYYJJJ}
  \item{size}{Each layer contains 113 rows and 105 columns}
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
ex.dem.navarre
}
\description{
Geographically projected \code{RasterBrick} object with the DEM values for
the region of Navarre, to use as example in \code{genSmoothingIMA}
}
\keyword{datasets}
