% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex.navarre}
\alias{ex.navarre}
\title{Geographically projected spatial polygon of Navarre in Spain to be used in examples}
\description{
\code{SpatialPolygonsDataFrame} containing the region of Navarre projected in longitude/latitude.
}
\keyword{data}
