% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex.ndvi.navarre}
\alias{ex.ndvi.navarre}
\title{NDVI of the tile containing Navarre in Spain}
\format{The \code{RasterBrick} contains 6 time-periods:
\describe{
  \item{name}{the name of each layer contains the date of the period in julian format as "\code{YYYYJJJ}"}.
  \item{size}{each layer contains 113 rows and 105 columns}.
}}
\usage{
ex.ndvi.navarre
}
\description{
Geographically projected \code{RasterBrick} object with the NDVI values for the region of Navarre.
}
\keyword{datasets}
