% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genFillingIMA.R
\name{genFillingIMA}
\alias{genFillingIMA}
\title{Fills the gaps in a time series of satellite images}
\usage{
genFillingIMA(imgTS, Img2Fill = NULL, aFilter = c(0.05, 0.95),
  fact = 5, nPeriods = 3, nYears = 1, fun = mean, factSE = 8,
  snow.mode = FALSE, predictSE = FALSE, out.name = "outname", ...)
}
\arguments{
\item{imgTS}{the time series of images to be filled}

\item{Img2Fill}{a vector defining the images to fill}

\item{aFilter}{two element vector defining the quantiles to filter the anomaly. Ex. c(0.05,0.95)}

\item{fact}{an aggregation factor to be used to reduce the anomalies before the interpolation}

\item{nPeriods}{number of periods used in the temporal window}

\item{nYears}{number of years used in the temporal window}

\item{fun}{function used to calculate the mean image, \code{mean}, \code{median} are acceptable functions}

\item{factSE}{the fact used in the standard error prediction to reduce the processing time}

\item{snow.mode}{a flag to use \code{raster} package parallelizing options for reconstructing (requires \code{snow} package)}

\item{predictSE}{a flag to calculate the standard error}

\item{out.name}{the name of the output images}

\item{...}{accepts \code{AppRoot} the path where the filled data will be saved or/and other argument for function nestering}
}
\value{
\code{RasterStack} object with the smoothed time series
}
\description{
\code{genFillingIMA} is the implementation of a spatio-temporal method called Image Mean Anomaly
for gap filling published in (\url{http://dx.doi.org/10.1109/TGRS.2019.2904193})
}
\details{
The time series is decomposed into a mean and a residual (anomaly)
components. Instead of smoothing images, the procedure applies
the smoothing algorithm over the anomaly.

The arguments configure the smoothing procedure:
\code{Img2Fill} identifies the images to filled, \code{nPeriods} defines
the spatio temporal neighbourhoood considered when smoothing
a pixel, and \code{aggfact} sets the level of spatial aggregation.
See more information in \url{http://dx.doi.org/10.1109/TGRS.2019.2904193}.
}
\examples{
# load an example of NDVI time series in Navarre
data(ex.ndvi.navarre)
# plot data example
genPlotGIS(ex.ndvi.navarre)

# fill the gaps
ndvi.filled<-genFillingIMA(ex.ndvi.navarre,
                           Img2Fill=c(1,2))
# Show the filled images
genPlotGIS(ndvi.filled)
# plot comparison of the cloud and the filled images
ndvi.comp<-stack(ex.ndvi.navarre[[1]],ndvi.filled[[1]],
                 ex.ndvi.navarre[[2]],ndvi.filled[[2]])
genPlotGIS(ndvi.comp,layout=c(2,2))
}
