% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genGetDates.R
\name{genGetDates}
\alias{genGetDates}
\title{Return the date of the capure from the name of the image}
\usage{
genGetDates(str, ...)
}
\arguments{
\item{str}{\code{character} argument containing the date in julian format "\code{YYYYJJJ}".}

\item{...}{argument for function nestering.
\itemize{
  \item \code{date.format} modify the format of the date being returned.
}}
}
\description{
\code{genGetDates} extracts the date of one or several images when the name of the layer includes julian date in format "\code{YYYYJJJ}".
}
\details{
The function reads a date from a \code{character} class variable in year-julian (\code{YYYYJJJ}) format and returns an R date class variable.
}
\examples{
img <- matrix(1:16, ncol = 4, byrow = TRUE)
r <- raster(img)
names(r) <- c("RandomImage_2018034")

spplot(r)
genGetDates(names(r), format = "\%Y\%j")

}
