% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genGetDates.R
\name{genGetDates}
\alias{genGetDates}
\title{Returns the capure date from name}
\usage{
genGetDates(str, ...)
}
\arguments{
\item{str}{string containing the date in julian format "\code{YYYYJJJ}".}

\item{...}{argument to allow function nestering.
\itemize{
  \item \code{date.format} modify the format of the date being returned.
}}
}
\description{
\code{genGetDates} extracts the date of one or several images when the date is included in julian date format "\code{YYYYJJJ}".
}
\details{
The function reads a date from a character type variable in year-julian format and returns an R date format variable.
}
\examples{
library(raster)

img <- matrix(1:16, ncol = 4, byrow = TRUE)
r<-raster(img)
names(r)<-c("RandomImage_2018034")

spplot(r)
genGetDates(names(r))

}
