% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genLimitRasterRange.R
\name{genLimitRasterRange}
\alias{genLimitRasterRange}
\title{Limits the maximum and minimum values of a raster image}
\usage{
genLimitRasterRange(r, ...)
}
\arguments{
\item{r}{raster type object}

\item{...}{accepts \code{mx} and \code{mn} as the maximun
value in the raster and the minimun value in the raster}
}
\description{
\code{genLimitRasterRange} limits the maximum and minimum values of a raster image to a given range
}
\details{
This is a generic function to limit the maximun and minimun values in a raster image.
}
\examples{
#generate random images
img <- matrix(1:16, ncol = 4, byrow = TRUE)
r<-raster(img)
#asign the limit of the data in the raster stack
r2<-genLimitRasterRange(r,mn=4,mx=10)
#plot limited data
spplot(stack(r,r2))
}
