% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genLimitRasterRange.R
\name{genLimitRasterRange}
\alias{genLimitRasterRange}
\title{Limits the maximum and the minimum values of a raster image}
\usage{
genLimitRasterRange(r, mx = NULL, mn = NULL, rm.values = FALSE)
}
\arguments{
\item{r}{\code{Raster}* type object.}

\item{mx}{maximun value in the \code{raster}.}

\item{mn}{minimun value in the \code{raster}.}

\item{rm.values}{logical argument. If \code{FALSE} \code{mx} and  \code{mn} values are assigned to the 
higher values than \code{mx} and lower values than \code{mn} respectively. If \code{TRUE}, values out of 
\code{mn}-\code{mx} are removed, and \code{NA} assigned.}
}
\description{
\code{genLimitRasterRange} limits the maximum and the minimum values of a raster image to a given range.
}
\details{
This is a generic function to limit the maximun and the minimun values in a \code{raster} image.
}
\examples{
# generate random images
img <- matrix(1:16, ncol = 4, byrow = TRUE)
r <- raster(img)
# assign the limit of the data in the raster stack
r2 <- genLimitRasterRange(r, mn = 4, mx = 10)
r3 <- genLimitRasterRange(r, mn = 4, mx = 10, rm.values = TRUE)
# plot limited data
spplot(stack(r, r2 ,r3))
}
