% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genMosaicList.R
\name{genMosaicList}
\alias{genMosaicList}
\title{Mosaics a list of raster images}
\usage{
genMosaicList(imageList, fun = "max", showWarnings = T)
}
\arguments{
\item{imageList}{list of raster images}

\item{fun}{the function applied to overlapping pixels}

\item{showWarnings}{\code{TRUE/FALSE} argument to print or omit warnnings}
}
\description{
\code{genMosaicList} makes a single mosaic from a list of raster images
}
\details{
This is a helper function used by other functions in this package. It combines
a list of raster images with different geolocation. If the images overlap, by
default the application selects the maximum value
}
\examples{
#create simulated rasters
img <- matrix(1:16, ncol = 4, byrow = TRUE)
r1<-raster(img)
r2<-r1
r3<-r1
extent(r2)<-extent(1,2,1,2)
extent(r3)<-extent(1,2,0,1)
imageList<-list(r1,r2,r3)
#mosaic simulated rasters
mr<-genMosaicList(imageList)
spplot(mr)
}
