% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSaveTSRData.R
\name{genSaveTSRData}
\alias{genSaveTSRData}
\title{Import time series of images and saves as RData}
\usage{
genSaveTSRData(src, ts.name = "TS.Name", startDate = NULL,
  endDate = NULL, dextent = FALSE, recursive = FALSE, ...)
}
\arguments{
\item{src}{path to the folder where the time series of images is located.}

\item{ts.name}{the name of the variable containing the time series in R.}

\item{startDate}{starting date of the image time series in \code{Date} class. For instance, using any format from \code{as.Date} function.}

\item{endDate}{ending date of the image time series in \code{Date} class. For instance, using any format from \code{as.Date} function.}

\item{dextent}{creates the \code{RasterStack} to the maximun extent from different extent GTiff images.}

\item{recursive}{logical argument. If \code{TRUE} reads folders recursively, searching for images in GTiff image format.}

\item{...}{argument for function nestering.
\itemize{
  \item \code{AppRoot} the path where the RData will be saved.
}}
}
\description{
\code{genSaveTSRData} imports the time series of (GTiff) satellite images into R from a folder, 
for creating an RData.
}
\details{
The function reads all the images inside the folder specified in \code{src}. The images must be GTiff files.
 The \code{src} can take the path created by other functions of this package, such as \code{\link{senMosaic}},
 \code{\link{modMosaic}}, \code{\link{senFolderToVar}}, etc. The images are imported into R to build a \code{RasterStack} that
 is loaded into the global environment. The name of the \code{RasterStack} is specified
 in \code{ts.name}. The \code{RasterStack} is saved into an RData file in the \code{AppRoot} directory.
}
\examples{
\dontrun{
# load a spatial polygon object of Navarre
data(ex.navarre)
# set the download folder
s.start <- Sys.time()
src <- "Path_for_downloading_folder"
# download the images
modDownload(product = "MOD09GA",
            startDate = as.Date("30-07-2018", "\%d-\%m-\%Y"),
            endDate = as.Date("06-08-2018", "\%d-\%m-\%Y"),
            username = "username",
            password = "password",
            AppRoot = src,
            extract.tif = TRUE,
            collection = 6,
            extent = ex.navarre)
# set folder path where MOD09GA images will be saved
src1 <- file.path(src,"MOD09GA")
# set the tif folder path
tif.src <- file.path(src1,"tif")
# mosaic and cut navarre region
modMosaic(tif.src,
          AppRoot = src1,
          out.name = "Navarre",
          extent = ex.navarre)
# change src to navarre folder
src2 <- file.path(src1,"Navarre")
# calculate NDVI from navarre folder
modFolderToVar(src2,
               fun = varNDVI,
               AppRoot = dirname(src2),
               overwrite = TRUE)
# change src TS_sample
src3 <- file.path(dirname(src2),"NDVI")
# create the Rdata
genSaveTSRData(src3, ts.name = "ModisNDVI", AppRoot = src)
s.end <- Sys.time()
s.end - s.start
}
}
