% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGISToolsOpts.R
\name{getRGISToolsOpt}
\alias{getRGISToolsOpt}
\title{Gets the current value of an RGISTools configuration variable}
\usage{
getRGISToolsOpt(opt, env = optEnv)
}
\arguments{
\item{opt}{the name of the option to get.}

\item{env}{the environment where the \code{RGISTools} option are saved.}
}
\description{
\code{getRGISToolsOpt} gets the current value of an \code{RGISTools} option.
}
\examples{
# list avaliable options names
showRGISToolsOpt()
# list the asignation of Sentinel bands
getRGISToolsOpt("SEN2BANDS")
# list the asignation of Landsat-8 bands
getRGISToolsOpt("LS8BANDS")
# list the asignation of Modis 09 bands
getRGISToolsOpt("MOD09BANDS")
}
