% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGISToolsOpts.R
\name{getRGISToolsOpt}
\alias{getRGISToolsOpt}
\title{gets the current value of an RGISTools option}
\usage{
getRGISToolsOpt(opt, env = optEnv)
}
\arguments{
\item{opt}{the name of the option to get}

\item{env}{the environment where the RGISTools option are saved}
}
\description{
gets the current value of an RGISTools option
}
\examples{
getRGISToolsOpt("SEN2BANDS")
getRGISToolsOpt("LS8BANDS")
}
