% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls7LoadMetadata.R
\name{ls7LoadMetadata}
\alias{ls7LoadMetadata}
\title{Sinchronize Landsat-7 meta data file for image search.}
\usage{
ls7LoadMetadata(update = FALSE, verbose = TRUE, omit.question = TRUE,
  ...)
}
\arguments{
\item{update}{logical argument. If \code{TRUE}, updates the metadata file.}

\item{verbose}{logical argument. If \code{TRUE}, the function prints running stages and warnings.}

\item{omit.question}{logical argument. If \code{TRUE}, omits the question for ensuring the loading of the metadata file.}

\item{...}{argument for function nestering:
\itemize{
  \item \code{AppRoot} this option specifies the downloading/searching directory of the metadata file. 
}}
}
\description{
\code{\link{ls7LoadMetadata}} loads a data frame called "\code{.LS7MD}"
 with the names of the Landsat-7 images and their metadata. The metadata provides
 auxiliary information regarding Landsat-7 repository, such as image quality, acquisition
 data, cloud cover, etc. You can find a description of the metadata at
 \href{https://www.usgs.gov/land-resources/nli/landsat/bulk-metadata-service}{bulk-metadata-service}.
}
\details{
All captures done by Landsat-7 are cataloged in a unique csv file. The size of the file
might be larger than 360MB. Therefore, the process of downloading and importing into R may take several
minutes (15 minutes in a Intel Core i7-4790, 16Gb of RAM and Hard Drive Device roughly). The function creates an
RData file with the csv metadata. Thus, every time \code{\link{ls7LoadMetadata}} is called,
this function loads the existing RData in the Approot. This is intended to reduce the loading
time of metadata in the future.
}
\examples{
\dontrun{
# creates a MetaData folder and downloads the csv on working directory
ls7LoadMetadata()

# creates a MetaData folder and downloads the csv on "Path_for_downloading_folder" directory
ls7LoadMetadata(AppRoot = "Path_for_downloading_folder")

# Force renew existing meta data csv
ls7LoadMetadata(update = TRUE)
# if .LS7MD is already loaded you can see print its data
ls(all.names = TRUE)
head(.LS7MD)
}
}
