% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls8FolderToVar.R
\name{ls8FolderToVar}
\alias{ls8FolderToVar}
\title{Computes derived variables from Landsat-8 multispectral bands}
\usage{
ls8FolderToVar(src, fun, getStack = FALSE, overwrite = FALSE, ...)
}
\arguments{
\item{src}{path to the folder with the Landsat multispectral image.}

\item{fun}{is a function defined by the package for computing indexes.
All functions in the package starting with three characters
'var' are acceptable functions. Custom functions can be also
implemented \code{var} are acceptable functions.}

\item{getStack}{logical argument. If \code{TRUE}, returns the 
time-series of images as a \code{RasterStack}, otherwise as Hard Drive Devide (HDD).}

\item{overwrite}{logical argument. If \code{TRUE} overwrites the 
existing images with the same name.}

\item{...}{argument to allow function nestering
\itemize{
  \item \code{AppRoot} the directory of the resulting time series
}}
}
\description{
\code{ls8FolderToVar} calculates an index using the bands from Landsat-8 multispectral
images. The images are specified by a path to the storing folder (resulting from
the \code{\link{lsMosaic}} function). The function returns a \code{RasterStack} with the time-series of the index.
}
\details{
The function requires to define \code{src} and \code{fun} attributes. \code{src} defines the path to
the result of \code{\link{lsMosaic}}, with all bands of Landsat-7 for a region of interest. \code{fun} defines
the variable of interest using any of the functions in the packages starting with \code{var} (\code{\link{varNDVI}},
 \code{\link{varEVI}}, ...)
}
\examples{
\dontrun{
# load a spatial polygon object of navarre for the example
data(ex.navarre)
# asign the folder where the example will be run
src <- "Path_for_downloading_folder"
# download Landsat-8 images
lsDownload(satellite = "ls8",
           username = "username",
           password = "password",
           startDate = as.Date("01-01-2018","\%d-\%m-\%Y"),
           endDate = as.Date("20-01-2018","\%d-\%m-\%Y"),
           extent = ex.navarre,
           untarDir = "untar",
           AppRoot = src)
# asign the folder with the Landsat-8 images untared
tif.src <- file.path(src, "untar")
# mosaic the Landsat-8 images
lsMosaic(tif.src,
         AppRoot = src,
         out.name = "Navarre")
# asign src as the path to mosaiced folder
src2 <- file.path(src, "Navarre")
# generate NDVI images of Navarre
ls8FolderToVar(src2,
               fun = varNDVI,
               AppRoot = src,
               overwrite = T)
}
}
