% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls8LoadMetadata.R
\name{ls8LoadMetadata}
\alias{ls8LoadMetadata}
\title{Sinchronizes landsat 8 meta data file, for image search.}
\usage{
ls8LoadMetadata(update = F, verbose = T, omit.question = F, ...)
}
\arguments{
\item{update}{TRUE/FALSE argument to force update of metadata file}

\item{verbose}{TRUE/FALSE argument to print all steps of metadata download process}

\item{omit.question}{omit the question for ensure the loading of the metadata}

\item{...}{accepts \code{AppRoot} as root directory where meta data file will be saved
or/and other argument for function nestering}
}
\description{
\code{ls8MetaDataFile} loads a data frame called \code{.LS8MD}
 with the names of the Landsat-8 images and their metadata.The metadata provides
 auxiliary information regarding Landsat-8 images repository such as image quality, acquisition
 data, cloud cover, etc. You can find a description of the metadata at
 \url{https://www.usgs.gov/land-resources/nli/landsat/bulk-metadata-service}
}
\details{
All captures done by Landsat-8 are cataloged in a unique csv file. The size of the file
might be larger than 210MB. Therefore, the process of downloading and importing to R may take several
minutes (around 7 minutes in a Intel Core i7-4790, 16Gb of RAM and Hard Drive storage). The function creates an
RData file with the metadata csv. Thus, every time \code{ls8MetaDataFile} is called,
this function loads the existing RData in the Approot. This is intended to reduce the loading
time of metadata in the future.
}
\examples{
\dontrun{
#creates a MetaData folder and downloads the csv on working directory
ls8LoadMetadata()

#creates a MetaData folder and downloads the csv on "C:/LandsatDownload" directory
ls8LoadMetadata(AppRoot="C:/LandsatDownload")

Force renew existing meta data csv
ls8LoadMetadata(update=TRUE)
}
}
