% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsCloudMask.R
\name{lsCloudMask}
\alias{lsCloudMask}
\title{Creates clouds layers for Landsat images}
\usage{
lsCloudMask(src, sensitivity = 2800, overwrite = FALSE, verbose = F,
  ...)
}
\arguments{
\item{src}{the path to the folder where the Landsat multispectral captures are stored.}

\item{sensitivity}{numeric argument. defines how sensitive is the method detecting the clouds. 0-8000 are
valid values.}

\item{overwrite}{logical argument. If \code{TRUE} overwrites the existing images with the same name.}

\item{verbose}{logical argument. If \code{TRUE} the function prints running stages and warnings.}

\item{...}{argument to allow function nestering:
\itemize{
  \item \code{AppRoot} the directory where the extracted images should be located
}}
}
\description{
\code{lsCloudMask} creates clouds layers derived from \code{BQA} band from Landsat-7 or Landsat-8 captures.
}
\examples{
\dontrun{
lsCloudMask(src="D:/Downscaling/Landsat8/untar",
            overwrite=T,
            sensitivity=98)
}
}
