% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsCloudMask.R
\name{lsCloudMask}
\alias{lsCloudMask}
\title{Creates clouds layers for Landsat images}
\usage{
lsCloudMask(src, sensitivity = 2800, overwrite = FALSE, verbose = F,
  ...)
}
\arguments{
\item{src}{the path to the folder where the Landsat multispectral captures are stored.}

\item{sensitivity}{numeric argument. Defines how sensitive is the method detecting the clouds. 0-8000 are
valid values.}

\item{overwrite}{logical argument. If \code{TRUE} overwrites the existing images with the same name.}

\item{verbose}{logical argument. If \code{TRUE} the function prints running stages and warnings.}

\item{...}{argument to allow function nestering:
\itemize{
  \item \code{AppRoot} the directory where the extracted images should be located
}}
}
\description{
\code{lsCloudMask} creates layers of clouds derived from \code{BQA} band from Landsat-7 or Landsat-8 captures. 
The function run over untar images downloaded by \code{\link{lsDownSearch}} or \code{\link{lsDownload}}, and 
creates a new band designated with the label \code{CLD}.
}
\examples{
\dontrun{
# load a spatial polygon object of Navarre
data(ex.navarre)
src <- "Path_for_downloading_folder"

# Search and download the images from Landsat-8 comprised between
# 01-01-2018 and 20-01-2018 for the region of Navarre
lsDownload(satellite = "ls8",
           username = "username",
           password = "password",
           startDate = as.Date("01-01-2018", "\%d-\%m-\%Y"),
           endDate = as.Date("20-01-2018", "\%d-\%m-\%Y"),
           extent = ex.navarre,
           untar = TRUE,
           AppRoot = src)
           
# generate the path where the GTiff images are located
src.untar <- file.path(src,"Landsat8","untar")
# calculate the cloud mask from QC layer
lsCloudMask(src=src.untar,
            overwrite=TRUE)
            
# generate the path where the mosacked images will be saved
src.ls8 <- dirname(src.untar)
lsMosaic(src = src.untar,
         AppRoot = src.ls8,
         out.name = "Navarre",
         extent = ex.navarre,
         bandFilter = c("CLD"),
         gutils = TRUE, # using gdalUtils
         overwrite = TRUE) # overwrite
         
# generate the path where mosaicked images are located
src.navarre <- file.path(src,"Landsat8","Navarre")
# load the B1 layer and calculated CLD layer
tiles.path <- list.files(src.navarre,
                         full.names = TRUE,
                         recursive = TRUE,
                         pattern = "\\\\.tif$")
cloud.tiles <- tiles.path[grepl("CLD",tiles.path)]
b1.tiles <- tiles.path[grepl("B1.tif",tiles.path)]
cloud.tiles.ras <- stack(cloud.tiles)
b1.tiles.ras <- stack(b1.tiles)

# calculate cloud free b1 layers
b1.cloud.free <- b1.tiles.ras * cloud.tiles.ras
spplot(b1.cloud.free)
}
}
