% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsDownSearch.R
\name{lsDownSearch}
\alias{lsDownSearch}
\title{Download a time series of satellite images from Landsat 7-8}
\usage{
lsDownSearch(searchres, username = NULL, password = NULL, lvl = 1,
  products = c("sr", "source_metadata"), verbose = FALSE,
  raw.rm = FALSE, untar = FALSE, overwrite = FALSE, n.attempts = 5,
  ...)
}
\arguments{
\item{searchres}{the results from \code{ls7Search} or \code{ls8Search}.}

\item{username}{login credentials to access the USGS EROS web service.}

\item{password}{login credentials to access the USGS EROS web service.}

\item{lvl}{flag to specify Landsat product level wanted. Default value, 1.}

\item{products}{character vector with the avaliable products for Landsat level 2. By default \code{c("sr","source_metadata")}.}

\item{verbose}{logical argument. If \code{TRUE}, the function prints running stages and warnings.}

\item{raw.rm}{logical argument. If \code{TRUE}, removes the raw images.}

\item{untar}{logical argument. If \code{TRUE}, untars downloaded images.}

\item{overwrite}{logical argument. If \code{TRUE}, overwrites the existing images with the same name.}

\item{...}{argument for function nestering accepts:
\itemize{
  \item \code{AppRoot} the directory to save the downloaded images.
}}

\item{cookies.file}{File path for saving the cookies that are used in the download process.}
}
\description{
\code{\link{lsDownSearch}} downloads the list of images provided by \code{\link{ls7Search}} and \code{\link{ls8Search}} functions.
The images are saved as ‘GTiff’ files in the \code{AppRoot} directory.
}
\details{
This function is used for downloading Landsat images. It downloads all the list of images 
provided in the output data frame of the search function (\code{\link{ls7Search}} or \code{\link{ls8Search}}).
Image download requires USGS login account. \href{https://ers.cr.usgs.gov/register/}{Get your credentials}.

The image files from the USGS EROS web service are compressed as ‘.tar.gz’ files. \code{\link{lsDownSearch}} decompresses the
images and obtains the corresponding ‘.tif’ files. The ‘.tif’ files are saved in the
\code{AppRoot} directory. When \code{untar} is defined the function untars the images in "untar" folder.
This replicates the images in compresses version as 'tar.gz' and uncompresses version in "untar" folder. To save space in the disk
\code{raw.rm = TRUE} can be defined, and \code{\link{lsDownSearch}} will remove the ‘tar.gz’ files.
If \code{raw.rm = FALSE}, the original files remain, which might be useful to have access to the original files
in the future and avoid further downloads. By default, \code{\link{lsDownSearch}} saves the images in (...), in the \code{AppRoot}
directory. To change this setting, provide \code{AppRoot = "the full path as an argument"}.
}
\examples{
\dontrun{
#' # load a spatial polygon object of Navarre
data(ex.navarre)

src <- "Path_for_downloading_folder"
# Search and download the images from Landsat-8 between
# 2011 and 2013 in the region of Navarre
search.res <- ls8Search(startDate = as.Date("01-01-2018", "\%d-\%m-\%Y"),
                        endDate = as.Date("20-01-2018", "\%d-\%m-\%Y"),
                        extent = ex.navarre,
                        browseAvaliable = "Y",
                        AppRoot = src)

# download 1 image
lsDownSearch(searchres = search.res[1,], 
             username = "user", 
             password = "pass", 
             untar = TRUE, 
             raw.rm = TRUE)
# download 10 images
lsDownSearch(searchres = search.res[1:4,], 
             username = "user", 
             password = "pass", 
             untar = TRUE, 
             raw.rm = TRUE)
# download all the images
lsDownSearch(searchres = search.res, 
             username = "user", 
             password = "pass", 
             untar = TRUE, 
             raw.rm = TRUE)

# Search and download the images from Landsat-7 between
# 2011 and 2013 in the region of Navarre
src <- "Path_for_downloading_folder"
search.res <- ls7Search(startDate = as.Date("01-01-2018", "\%d-\%m-\%Y"),
                        endDate = as.Date("20-01-2018", "\%d-\%m-\%Y"),
                        extent = ex.navarre,
                        browseAvaliable = "Y",
                        AppRoot=src)
# download 1 image
lsDownSearch(searchres = search.res[1,], 
             username = "user", 
             password = "pass", 
             untar = TRUE, 
             raw.rm = TRUE,
             AppRoot=src)
# download 10 images
lsDownSearch(searchres = search.res[1:10,], 
             username = "user", 
             password = "pass", 
             untar = TRUE, 
             raw.rm = TRUE,
             AppRoot=src)
# download all the images
lsDownSearch(searchres = search.res, 
             username = "user", 
             password = "pass", 
             untar = TRUE, 
             raw.rm = TRUE,
             AppRoot=src)

# removes metadata data frame to free memory
lsRemoveMetadata()

files <- list.files("./", 
                    pattern = "\\\\.TIF$", 
                    full.names = TRUE, 
                    recursive = TRUE)[6,5,4]
files.stack <- stack(files)
qrange <- c(0.001, 0.999)
imagen <- varRGB(files.stack.raster[[1]], 
                 files.stack.raster[[2]],
                 files.stack.raster[[3]],
                 qrange)
plotRGB(imagen)
}
}
