% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genMissionGetDates.R
\name{lsGetDates}
\alias{lsGetDates}
\title{Returns the capturing date of a Landsat-7 or Landsat-8 images}
\usage{
lsGetDates(str, ...)
}
\arguments{
\item{str}{a string or a list of strings defining the file path(s) or file name(s).}

\item{...}{argument to allow function nestering:
\itemize{
  \item \code{date.format} modify the format of the date being returned.
}}
}
\description{
\code{lsGetDates} extracts the capturing date of one or several Landsat images,
given their file paths or names. The function returns a date class object.
}
\details{
The function works with names or file paths having either of the two file extensions,
 \code{.tar.gz} or \code{.tif} files. The function accepts more than on file path or file names.
 If so, they should be provided as a list. Dates are returned as '\code{YYYY-mm-dd}' by default.
 If another format is required, it can be modified through the argument \code{date.format}.
}
\examples{
# example of getting the date from the name of a Landsat-8 image
str <- "LC82000312017010LGN01_B1.TIF"
dt <- lsGetDates(str)
print(dt)
print(format(df, "\%Y\%j"))

# example of getting the date from the name of a Landsat-7 and 8 image
str <- c("LE72330822017009ASN01", "LC82000312017010LGN01_B1.TIF")
dt <- lsGetDates(str)
print(dt)

}
