% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsGetPathRow.R
\name{lsGetPathRow}
\alias{lsGetPathRow}
\title{Returns the pathrow of Modis tile in "hXXvYY" string format}
\usage{
lsGetPathRow(str)
}
\arguments{
\item{str}{the full path or official image name of the Lansat 7 or 8 image from which the path and row will be extracted}
}
\value{
an string with the path and row of the image in "ppprrr" format
}
\description{
\code{lsGetPathRow} reads the official name of a Lansat 7 or 8 image and returns the path and row
}
\examples{
#example of getting date from Landsat-8 image name
str<-c("LE72000302011066ASN00",
       "LE72000302011066ASN00")
pr<-lsGetPathRow(str)
print(pr)

}
