% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsRemoveMetadata.R
\name{lsRemoveMetadata}
\alias{lsRemoveMetadata}
\title{Removes from R environment LS7 and LS8 metadata files}
\usage{
lsRemoveMetadata()
}
\description{
\code{lsRemoveMetadata} frees the memory used by metadata data frame loaded in R enviroment.
}
\details{
If the metadata file has been loaded by any landsat search function or using load
metadata funcion, \code{lsRemoveMetadata} removes anoy of both dataframes \code{.LS8MD} or \code{.LS7MD}.
}
\examples{
\dontrun{
#creates a MetaData folder and downloads the csv on working directory
ls8LoadMetadata()
lsRemoveMetadata()
}
}
