% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modCloudMask.R
\name{modCloudMask}
\alias{modCloudMask}
\title{Create cloud masks for MODIS images}
\usage{
modCloudMask(src, AppRoot, out.name, overwrite = FALSE, ...)
}
\arguments{
\item{src}{the path to the folder with the MODIS with \code{state_1km} images.}

\item{AppRoot}{the directory where cloud masks are saved.}

\item{out.name}{the name of the folder that stores the outputs. If the arguemnt is defined 
as "outname", the name of the folder will be named "outname_CloudMask". 
If the arguemnt is not defined the folder will be named as "CloudMask".}

\item{overwrite}{logical argument. If \code{TRUE}, overwrites the existing
images with the same name.}

\item{...}{arguments for nested functions.
\itemize{
  \item \code{dates} a vector with the dates being considered
  for creating cloud mask. This argument is optional.
}}
}
\description{
\code{modCloudMask} creates cloud masks derived from the "\code{MOD35_L2}" 
product.
}
\details{
This function, downloads and processes the "\code{MOD35_L2}" products to create
cloud masks composed of \code{NA}'s (cloud) and \code{1}'s (clear). The
resulting cloud mask layers need to be reprojected because resolution and
projection differences with other MODIS products. This function requires
`GDAL' and the `\code{gdalUtils}' library properly installed.
}
\examples{
\dontrun{
# load a spatial polygon object of Navarre
data(ex.navarre)
src <- paste0(tempdir(),"/Path_for_downloading_folder")
print(src)

# search and download images from MODIS between
# 01-01-2018 and 03-01-2018 for the region of Navarre
modDownSearch(product = "MOD09GA",
            startDate = as.Date("01-01-2017", "\%d-\%m-\%Y"),
            endDate = as.Date("03-01-2017", "\%d-\%m-\%Y"),
            username = "username",
            password = "password",
            AppRoot = src,
            extract.tif = TRUE,
            collection = 6,
            extent = ex.navarre)
            
# assign src1 as the output folder for modMosaic
src.mod <- file.path(src, "Modis")
src.tiles <- file.path(src.mod, "MOD09GA")
tif.src.tiles <- file.path(src.tiles, "tif")
# mosaic the MODIS images
modMosaic(tif.src.tiles, # the input folder 
          AppRoot = src.tiles, # the output folder 
          out.name = "Navarre", # creates Navarre folder in AppRoot
          gutils = TRUE,
          extent = ex.navarre)


src.tiles.navarre <- file.path(src.tiles, "Navarre")
# generate the cloud masks      
modCloudMask(src = src.tiles.navarre,
             AppRoot = src.tiles,
             overwrite = TRUE)
             
src.cloud <- file.path(src.tiles,"CloudMask")
src.cloud.stack <- stack(list.files(src.cloud, full.names=TRUE, pattern="CLD"))

# select b01
navarre.img <- stack(list.files(tif.src.tiles, 
                                full.names=TRUE, 
                                recursive = TRUE, 
                                pattern="b01_1"))

# project to 500m
src.cloud.stack <- projectRaster(src.cloud.stack,navarre.img)

# plot the cloud free b01 layer
spplot(navarre.img*src.cloud.stack)
}
}
