% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modDownload.R
\name{modDownload}
\alias{modDownload}
\title{Search and downloads Modis images in structured folder}
\usage{
modDownload(product, startDate, endDate, username, password,
  collection = 6, verbose = FALSE, hdfdir = "hdf", tiffdir = "tif",
  ...)
}
\arguments{
\item{product}{Modis product type.}

\item{startDate}{start date of time series.}

\item{endDate}{end date of time series.}

\item{username}{EarthData username.}

\item{password}{EarthData password.}

\item{collection}{Modis collection.}

\item{verbose}{logical argument. If \code{TRUE} the function prints running stages and warnings.}

\item{hdfdir}{folder name for hdf download.}

\item{tiffdir}{folder name for tif format images.}

\item{...}{argument for function nestering:
\itemize{
  \item \code{lonlat} A vector or a polygon with the coordinates of
the point or region of interest in latitude/longitude format.
This argument is mandatory if polygon or extent is not defined.
  \item \code{extent} \code{Extent}, \code{Raster*}, \code{SpatialPolygons*}, \code{SpatialLines*} or \code{SpatialPoints*} object are acceptable formats
as long as coordinates are in latitude/longitude format.
This argument is mandatory if \code{polygon} or \code{lonlat} is not defined.
  \item \code{polygon} A list of vectors defining the points of the polygon in latitude/longitude format
This argument is mandatory if \code{lonlat} or extent is not defined.
  \item \code{AppRoot} the directory to save the resulting time series
}}
}
\description{
\code{modDownload} searches and downloads the Modis products on the  NASA Common Metadata
Repository to find those which are relevant for a particular location and date interval.
}
\details{
\code{modDownload} NASA Common Metadata Repository (CMR) powered api \url{https://lpdaacsvc.cr.usgs.gov/services/inventory}
The catalogue of  Modis products with their short names and other
information can be found at: \url{https://modis.gsfc.nasa.gov/data/dataprod/}
For further information on collections, please visit Modis website at: \url{https://modis-atmos.gsfc.nasa.gov/collections/overview}
By the time the RGISTools package is released, NASA carries out the maintenance of the Modis website on Wednesdays. Therefore, an error
may occur when trying to connect with their server during this day of the week.
}
\examples{
\dontrun{
data(ex.navarre)
modDownload(product="MOD09GA",
           startDate=as.Date("01-01-2018","\%d-\%m-\%Y"),
           endDate=as.Date("03-01-2018","\%d-\%m-\%Y"),
           username="username",
           password="password",
           AppRoot="Path_for_downloading_folder",
           hdfdir="hdf",
           tiffdir="tif",
           collection=6,
           extent=ex.navarre)
}
}
