% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modDownload.R
\name{modDownload}
\alias{modDownload}
\title{Search and downloads Modis images in structured folder}
\usage{
modDownload(product, startDate, endDate, username, password,
  collection = 6, verbose = FALSE, extract.tif = FALSE, ...)
}
\arguments{
\item{product}{a \code{character} argument with the name of Modis product type.}

\item{startDate}{starting date of the image time series in \code{Date} class. For instance, using any format from \code{as.Date} function.}

\item{endDate}{ending date of the image time series in \code{Date} class. For instance, using any format from \code{as.Date} function.}

\item{username}{EarthData username.}

\item{password}{EarthData password.}

\item{collection}{Modis collection.}

\item{verbose}{logical argument. If \code{TRUE} the function prints running stages and warnings.}

\item{extract.tif}{logical argument. If \code{TRUE} extracts as tif image format all the layers in a hdf image.}

\item{...}{argument for function nestering:
\itemize{
  \item \code{lonlat} A vector or a polygon with the coordinates of
the point or region of interest in longitude/latitude format.
This argument is mandatory if polygon or extent is not defined.
  \item \code{extent} \code{Extent}, \code{Raster*}, \code{SpatialPolygons*}, \code{SpatialLines*} or 
  \code{SpatialPoints*} object are acceptable formats as long as coordinates 
  are in longitude/latitude format. This argument is mandatory if \code{polygon} 
  or \code{lonlat} is not defined.
  \item \code{polygon} A list of vectors defining the points of the polygon in longitude/latitude format
This argument is mandatory if \code{lonlat} or extent is not defined.
  \item \code{AppRoot} the directory to save the resulting time series
}}
}
\description{
\code{modDownload} searches and downloads the Modis products on the 
\href{https://lpdaacsvc.cr.usgs.gov/services/inventory}{NASA Common Metadata Repository} to find images 
within a particular location and date interval.
}
\details{
\code{modDownload} uses \href{https://lpdaacsvc.cr.usgs.gov/services/inventory}{NASA Common Metadata Repository} (CMR) 
powered api to search satellite all ground products releses by NASA. The catalogue of Modis ground products with their 
short names and other information can be found at: \href{https://modis.gsfc.nasa.gov/data/dataprod/}{Modis data product info}.
For further information on collections, please visit \href{https://modis-atmos.gsfc.nasa.gov/collections/overview}{Modis website}.
By the time the \code{RGISTools} package is released, NASA carries out the maintenance of the Modis website on Wednesdays. 
Therefore, an error may occur when trying to connect with their server during this day of the week.
\href{https://urs.earthdata.nasa.gov/users/new}{Get your credentials}.
}
\examples{
\dontrun{
data(ex.navarre)
modDownload(product = "MOD09GA",
            startDate = as.Date("01-01-2018", "\%d-\%m-\%Y"),
            endDate = as.Date("03-01-2018", "\%d-\%m-\%Y"),
            username = "username",
            password = "password",
            AppRoot = "Path_for_downloading_folder",
            extract.tif = TRUE,
            collection = 6,
            extent = ex.navarre)
files <- list.files("./Path_for_downloading_folder/tif",
                    pattern = "\\\\.tif$",
                    full.names = TRUE,
                    recursive = TRUE)[1,4,3]
files.stack <- stack(files)
qrange <- c(0.001, 0.999)
imagen <- varRGB(files.stack.raster[[1]], 
                 files.stack.raster[[2]],
                 files.stack.raster[[3]],
                 qrange)
plotRGB(imagen)
}
}
