% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modDownload.R
\name{modDownload}
\alias{modDownload}
\title{Search and downloads Modis images in structured folder}
\usage{
modDownload(product, startDate, endDate, username, password,
  collection = 6, verbose = F, hdfdir = "hdf", tiffdir = "tif",
  ...)
}
\arguments{
\item{product}{modis product type}

\item{startDate}{Start date of time series}

\item{endDate}{End date of time series}

\item{username}{EarthData username}

\item{password}{EarthData password}

\item{collection}{Modis collection}

\item{verbose}{Flag for printing debugging information}

\item{hdfdir}{folder name for hdf download}

\item{tiffdir}{folder name for tif format images}

\item{...}{argument for function nestering
\itemize{
  \item \code{AppRoot} Aplication environment root path
  \item \code{latlon} Vectorial projected file
}}
}
\description{
\code{modDownload} searches and downloads the MODIS products on the  NASA Common Metadata
Repository to find those which are relevant for a particular location and date interval.
}
\details{
\code{modDownload} NASA Common Metadata Repository (CMR) powered api \url{https://lpdaacsvc.cr.usgs.gov/services/inventory}
The catalogue of  MODIS products with their short names and other
information can be found at: \url{https://modis.gsfc.nasa.gov/data/dataprod/}
For further information on collections, please visit MODIS website at: \url{https://modis-atmos.gsfc.nasa.gov/collections/overview}
By the time the RGISTools package is released, NASA carries out the maintenance of the MODIS website on Wednesdays. Therefore, an error
may occur when trying to connect with their server during this day of the week.
}
\examples{
\dontrun{
data(navarre)
modDownload(product="MOD09GA",
           startDate=as.Date("01-01-2018","\%d-\%m-\%Y"),
           endDate=as.Date("03-01-2018","\%d-\%m-\%Y"),
           username="rgistools",
           password="EspacialUPNA88",
           AppRoot="D:/TestEnvironment",
           hdfdir="hdf",
           tiffdir="tif",
           collection=6,
           extent=navarre)
}
}
