% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modDownSearch.R
\name{modDownSearch}
\alias{modDownSearch}
\title{Downloads Modis images from search function response}
\usage{
modDownSearch(searchres, username = NULL, password = NULL,
  overwrite = FALSE, ...)
}
\arguments{
\item{searchres}{output from \code{\link{modSearch}} function.}

\item{username}{login credentials to access the NASA’s EarthData web service.}

\item{password}{login credentials to access the NASA’s EarthData web service.}

\item{overwrite}{logical argument. If \code{TRUE} overwrites the existing images with the same name.}

\item{...}{argument to allow function nestering:
\itemize{
  \item \code{AppRoot} root directory where the images will be saved.
}}
}
\description{
\code{modDownSearch} downloads the images from a list of urls generated by the \code{modSearch} from NASA’s EartData plataform.
}
\details{
\code{modDownSearch} is able to download Modis Terra and Aqua products.
These products are published in the \href{https://earthdata.nasa.gov}{Earth Data Platform}. The platform
is supported by the Earth Observing System Data and Information System (EODIS)
and managed NASA’s Earth Science Data Systems (ESDS).
\code{modDownSearch} requires the credentianls from EarthData to access the NASA’s web data service.
\href{https://urs.earthdata.nasa.gov/users/new}{Get credentials}.
}
\examples{
\dontrun{
data(ex.navarre)
mList <- modSearch(product = "MYD13A2",
                   startDate = as.Date("01-01-2011", "\%d-\%m-\%Y"),
                   endDate = as.Date("31-12-2013", "\%d-\%m-\%Y"),
                   collection = 6,
                   extent = ex.navarre)
head(mList)
# download first image in mList
modDownSearch(mList[1], "user", "pass")
# download all images in mList
modDownSearch(mList, "user", "pass")
}
}
