% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modDownSearch.R
\name{modDownSearch}
\alias{modDownSearch}
\title{Downloads MODIS images from search function response}
\usage{
modDownSearch(searchres, username, password, overwrite = FALSE, ...)
}
\arguments{
\item{searchres}{url list made by \code{\link{modSearch}} function.}

\item{username}{login credentials to access the NASA’s EarthData web service.}

\item{password}{login credentials to access the NASA’s EarthData web service.}

\item{overwrite}{logical argument. If \code{TRUE} overwrites the existing images with the same name.}

\item{...}{argument to allow function nestering:
\itemize{
  \item \code{AppRoot}  Root directory where the images will be saved.
}}
}
\description{
\code{modDownSearch} downloads donwloads the images from a list of urls generated by the \code{modSearch} from NASA’s EartData plataform.
}
\details{
\code{modDownSearch} is able to download MODIS Terra and Aqua products.
These products are published in the Earth Data Platform (\url{https://earthdata.nasa.gov}). The platform
is supported by the Earth Observing System Data and Information System (EODIS)
and managed NASA’s Earth Science Data Systems (ESDS).
\code{modDownSearch} requires the credentianls from EarthData to access the NASA’s web data service.
You cansign up at \url{https://urs.earthdata.nasa.gov/users/new}.
}
\examples{
\dontrun{
data(ex.navarre)
mList<-modSearch(product="MYD13A2",
                 startDate=as.Date("01-01-2011","\%d-\%m-\%Y"),
                 endDate=as.Date("31-12-2013","\%d-\%m-\%Y"),
                 collection=6,
                 latlon=ex.navarre)
head(mList)
#download first image in mList
modDownSearch(mList[1],"user","pass")
#download all images in mList
modDownSearch(mList,"user","pass")
}
}
