% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modFolderToVar.R
\name{modFolderToVar}
\alias{modFolderToVar}
\title{Calculates an index from Modis multispectral bands}
\usage{
modFolderToVar(src, fun, getStack = FALSE, overwrite = FALSE, ...)
}
\arguments{
\item{src}{path to the folder with the Modis multispectral images.}

\item{fun}{is a function with the calculation of the index.
All functions in the package starting with three characters
'var' are acceptable functions. Custom functions can be also implemented.}

\item{getStack}{logical argument. If \code{TRUE}, returns the time-series as a raster or otherwise as Hard Drive Devide (HDD).}

\item{overwrite}{logical argument. If \code{TRUE} overwrites the existing images with the same name.}

\item{...}{argument to allow function nestering:
\itemize{
  \item \code{AppRoot} the directory of the resulting time series.
}}
}
\description{
\code{\link{modFolderToVar}} calculates an index using the bands from Modis multispectral images.
The images are specified by a path to the storing folder (resulting from the \code{\link{modMosaic}} function).
The function returns a \code{RasterStack} with the time series of the index.
}
\details{
The function requires the definition of src and fun arguments. The argument \code{src}
contains the path to the folder with the multispectral images. It can be easily
defined as the path resulting from \code{\link{modMosaic}}. The fun argument is a function with
the calculation of the index based on spectral bands. There are some pre-programmed
indices in \code{RGISTools}. Functions with the pre-programmed indexes start with var
(\code{\link{varNDVI}}, \code{\link{varEVI}}). The user can define its own functions.
}
\examples{
\dontrun{
# load a spatial polygon object of navarre for the example
data(ex.navarre)
# asign the folder where the example will be run
src<-"Path_for_downloading_folder"
# download Modis images
modDownload(product="MOD09GA",
            startDate=as.Date("01-01-2018","\%d-\%m-\%Y"),
            endDate=as.Date("03-01-2018","\%d-\%m-\%Y"),
            username="username",
            password="password",
            AppRoot=src,
            hdfdir="hdf",
            tiffdir="tif",
            collection=6,
            extent=ex.navarre)
# asign the folder with the Sentinel images untared
src<-file.path(src,"MOD09GA")
tif.src<-file.path(src,"tif")
#mosaic the Modis images
modMosaic(tif.src,
          AppRoot=src,
          out.name="Navarre")
# asign src as the path to mosaiced folder
src2<-file.path(src,"Navarre")
# generate NDVI images of Navarre
modFolderToVar(src2,
               fun=varEVI,
               AppRoot=src),
               overwrite = T)
}
}
