% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modGetPathRow.R
\name{modGetPathRow}
\alias{modGetPathRow}
\title{Returns the pathrow of Modis tile in "hXXvYY" string format}
\usage{
modGetPathRow(str)
}
\arguments{
\item{str}{the full path or official image name of the Modis image from which the path and row is extracted.}
}
\value{
an string with the path and row of the image in "\code{hXXvYY}" format.
}
\description{
\code{\link{modGetPathRow}} reads the official name of a Modis image and returns the path and row data
as Modis name convention of horizontal and vertical "\code{hXXvYY}".
}
\examples{
# example of getting date from Landsat-8 image name
str <- "MYD09GA.A2003136.h17v04.005.2008324054225"
pr <- modGetPathRow(str)
print(pr)

}
