% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modSearch.R
\name{modSearch}
\alias{modSearch}
\title{Search Modis Images on the NASA Common Metadata Repository}
\usage{
modSearch(product, startDate, endDate, collection = 6, resType = "url",
  verbose = FALSE, ...)
}
\arguments{
\item{product}{the short name of the Modis product.}

\item{startDate}{starting date of the image time series in \code{Date} class. For instance, using any format from \code{as.Date} function.}

\item{endDate}{ending date of the image time series in \code{Date} class. For instance, using any format from \code{as.Date} function.}

\item{collection}{Modis collection.}

\item{resType}{response type of the query (\code{browseurl} or \code{url}), by default the url where the images are located.}

\item{verbose}{logical argument. If \code{TRUE}, the function prints running stages and warnings.}

\item{...}{argument to allow function nestering:
\itemize{
  \item \code{lonlat} a vector or a polygon with the coordinates of
the point or region of interest in longitude/latitude format.
This argument is mandatory if polygon or extent is not defined.
  \item \code{extent} \code{Extent}, \code{Raster*}, \code{SpatialPolygons*}, \code{SpatialLines*} or \code{SpatialPoints*} object are acceptable formats
as long as coordinates are in longitude/latitude format.
This argument is mandatory if \code{polygon} or \code{lonlat} is not defined.
  \item \code{polygon} a list of vectors defining the points of the polygon in longitude/latitude format.
This argument is mandatory if \code{lonlat} or \code{extent} is not defined.
  \item \code{AppRoot} the directory to save the output time series.
}}
}
\description{
\code{modSearch} searches the Modis products on the  
\href{https://lpdaacsvc.cr.usgs.gov/services/inventory}{NASA Common Metadata Repository} to 
 find the avaliable images for a particular location and date interval.
The outut is an array of urls with the relevant images.
}
\details{
\code{modSearch} \href{https://lpdaacsvc.cr.usgs.gov/services/inventory}{NASA Common Metadata Repository} (CMR) powered api.
The catalogue of  Modis products with their short names and other
information can be found at: \href{https://modis.gsfc.nasa.gov/data/dataprod/}{Modis data product info}.
For further information on collections, please visit \href{https://modis-atmos.gsfc.nasa.gov/collections/overview}{Modis website}.
By the time the \code{RGISTools} package is released, NASA carries out the maintenance of the Modis website on Wednesdays. Therefore, an error
may occur when trying to connect with their server during this day of the week.
}
\examples{
\dontrun{
# load a spatial polygon object of Navarre in longitude/latitude coordinates
data(ex.navarre)
# Searching Modis MYD13A2 images between 2011 and 2013 by longitude/latitude
using a polygon class variable
img.list <- modSearch(product = "MYD13A2",
                      startDate = as.Date("01-01-2011", "\%d-\%m-\%Y"),
                      endDate = as.Date("31-12-2013", "\%d-\%m-\%Y"),
                      collection = 6,
                      extent = ex.navarre)
# Area of interest: defined based on longitude-latitude extent
# Searching Modis MYD13A2 images in the whole 2010 by longitude/latitude
# using a extent class variable defined by the user
aoi = extent(c(-2.49, -0.72, 41.91, 43.31))
my.imgs <- modSearch(product = "MYD13A2",
                     startDate = as.Date("01-01-2010", "\%d-\%m-\%Y"),
                     endDate = as.Date("31-12-2010", "\%d-\%m-\%Y"),
                     collection = 6,
                     extent = aoi)
head(my.imgs)
}
}
