% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senCloudMask.R
\name{senCloudMask}
\alias{senCloudMask}
\title{Creates clouds layers for sentinel-2 images}
\usage{
senCloudMask(src, img.res, sensitivity = 50, overwrite = FALSE, ...)
}
\arguments{
\item{src}{the path to the folder where the \code{S2MSI2A} product are stored.}

\item{img.res}{character vector argument. Defines the resolution used to create the cloud mask. ex c("10m", "20m", "30m").}

\item{sensitivity}{numeric argument. defines how sensitive is the method detecting the clouds. 0}

\item{overwrite}{logical argument. If \code{TRUE} overwrites the existing images with the same name.}

\item{...}{argument to allow function nestering:
\itemize{
  \item \code{AppRoot} the directory where the extracted images should be located
}}
}
\description{
\code{senCloudMask} creates clouds layers using \code{CLDPROB} band from \code{S2MSI2A} product.
}
\examples{
\dontrun{
senCloudMask(src="D:/Downscaling/Sentinel2_L2/Navarre",
             img.res="20m",
             overwrite=TRUE,
             AppRoot="D:/Downscaling/Sentinel2_L2",
             sensitivity=98)
}
}
