% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senCloudMask.R
\name{senCloudMask}
\alias{senCloudMask}
\title{Creates clouds layers for sentinel-2 images}
\usage{
senCloudMask(src, img.res, sensitivity = 50, overwrite = FALSE, ...)
}
\arguments{
\item{src}{the path to the folder where the \code{S2MSI2A} product are stored.}

\item{img.res}{character vector argument. Defines the resolution used to create the cloud mask. ex c("10m", "20m", "30m").}

\item{sensitivity}{numeric argument. defines how sensitive is the method detecting the clouds. 0}

\item{overwrite}{logical argument. If \code{TRUE} overwrites the existing images with the same name.}

\item{...}{argument to allow function nestering:
\itemize{
  \item \code{AppRoot} the directory where the extracted images should be located.
}}
}
\description{
\code{senCloudMask} creates clouds layers using \code{CLDPROB} band from \code{S2MSI2A} product.
}
\examples{
\dontrun{
# Download S2MSI1C products sensed by Sentinel - 2 
# satellite between the julian dates 210 and 218, 2018
data(ex.navarre)
src <- "Path_for_downloading_folder"
senDownload(startDate = as.Date("2018210", "\%Y\%j"),
            endDate = as.Date("2018218", "\%Y\%j"),
            platform = "Sentinel-2",
            extent = ex.navarre,
            product = "S2MSI1C",
            pathrow = c("R094"),
            username = "username",
            password = "password",
            unzip = TRUE,
            AppRoot = src)
src.unzip <- file.path(src, "unzip")
senMosaic(src.unzip,
          AppRoot = src,
          gutils = TRUE,
          out.name = "Navarre")
src.mosaic <- file.path(src, "Navarre")
src.cloud <- file.path(src, "CloudMask")
senCloudMask(src = src.mosaic,
             img.res = "20m",
             overwrite = TRUE,
             AppRoot = src.cloud,
             sensitivity=98)
}
}
