% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senFolderToVar.R
\name{senFolderToVar}
\alias{senFolderToVar}
\title{Calculates an index from Sentinel-2 multispectral bands}
\usage{
senFolderToVar(src, fun, getStack = FALSE, overwrite = FALSE, ...)
}
\arguments{
\item{src}{path to the folder with the Sentinel multispectral images.}

\item{fun}{function with the calculation of the index.
All functions in the package starting with three characters.
'var' are acceptable functions. Custom functions can be also implemented.}

\item{getStack}{logical argument. If \code{TRUE}, returns the time-series as a raster or otherwise as Hard Drive Devide (HDD).}

\item{overwrite}{logical argument. If \code{TRUE} overwrites the existing images with the same name.}

\item{...}{argument to allow function nestering:
\itemize{
  \item \code{AppRoot} the directory of the resulting time series.
}}
}
\description{
\code{\link{senFolderToVar}} calculates an index using the bands from Sentinel-2 multispectral
images. The images are specified by a path to the storing folder
(resulting from the \code{\link{senMosaic}} function). The function returns a
\code{rasterStack} with the time-series of the index.
}
\details{
The function requires the definition of src and fun arguments. The argument \code{src}
 contains the path to the folder with the multispectral
 images. It can be easily defined as the path resulting from \code{\link{senMosaic}}.
 The fun argument is a function with the calculation of the index based
 on spectral bands. There are some pre-programmed indexes in \code{RGISTools}.
 Functions with the pre-programmed indices start with var (\code{\link{varNDVI}}, \code{\link{varEVI}}).
 The user can define its own functions.
}
\examples{
\dontrun{
#load a spatial polygon object of navarre for the example
data(ex.navarre)
#asign the folder where the example will be run
src<-"Path_for_downloading_folder"
#download sentinel images
senDownload(startDate=as.Date("2018210","\%Y\%j"),
            endDate=as.Date("2018218","\%Y\%j"),
            platform="Sentinel-2",
            extent=ex.navarre,
            product="S2MSI1C",
            pathrow=c("R094"),
            username="username",
            password="password",
            AppRoot=src)
#asign the folder with the sentinel images unzipped
src.unzip<-file.path(src,"unzip")
#mosaic the sentinel images
senMosaic(src.unzip,
          AppRoot=src,
          gutils=T,
          out.name="Navarre")
#asign src as the path to mosaiced folder
src<-file.path(src,"Navarre")
#generate EVI images of Navarre
senFolderToVar(src,
               fun=varEVI,
               AppRoot=file.path(dirname(src)))
}
}
