% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genMissionGetDates.R
\name{senGetDates}
\alias{senGetDates}
\title{Returns the capturing date of Sentinel image}
\usage{
senGetDates(str, ...)
}
\arguments{
\item{str}{\code{character} class or a \code{vector} class containing the file path(s) or file name(s).}

\item{...}{argument to allow function nestering:
\itemize{
  \item \code{date.format} modify the format of the date being returned.
}}
}
\description{
\code{senGetDates} extracts the capturing date of one or several Sentinel
images, given their file paths or names. The function returns a date class.
}
\details{
The function works with names or file paths having either of the two file extensions,
 \code{.jp2} or \code{.tif} files. The function accepts more than on file path or file names.
 If so, they should be provided as a list. Dates are returned as '\code{YYYY-mm-dd}' by default.
 If another format is required, it can be modified through the argument \code{date.format}.
}
\examples{
# example of getting date from Sentinel2 image name
str <- c("S2A_MSIL1C_20170102T111442_N0204_R137_T30TWN_20170102T111441.SAFE",
       "S2A_OPER_PRD_MSIL1C_PDMC_20160308T090616_R094_V20160305T110109_20160305T110109")
dt <- senGetDates(str)
print(dt)
print(format(dt,"\%Y\%j"))
senGetDates(str, date.format = "\%Y\%j")

}
