% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senPreview.R
\name{senPreview}
\alias{senPreview}
\title{Allows Sentinel satellite images preview before its download}
\usage{
senPreview(searchres, username, password, n, size = NULL)
}
\arguments{
\item{searchres}{a data frame with  the results from a search of Landsat images provided by the function senSearch}

\item{username}{Scihub username}

\item{password}{Scihub password}

\item{n}{a number with the row corresponding to the image of interest in the search data frame}

\item{size}{a number specifying the size of the preview to be displayed. The number determines pixels numbe}
}
\description{
\code{senPreview} shows a preview of the n image from a set of search results
}
\details{
The functions shows a preview of an image resulting from a search in Scihub platform.
A search with \code{senSearch} has to be carried before proceeding with the preview.
The preview is downloaded from Scihub website, and the crediantials are needed.
Please, be aware that only some images have this feature.
}
\examples{
\dontrun{
data(navarre)
# perform the search query
searchres<-senSearch(startDate=as.Date("2018210","\%Y\%j"),
                     endDate=as.Date("2018218","\%Y\%j"),
                     platform="Sentinel-2",
                     intersects=navarre,
                     product="S2MSI1C",
                     username="rgistools",
                     password="EspacialUPNA88")
#Preview some images
senPreview(searchres,3,username="rgistools",password="EspacialUPNA88")
senPreview(searchres,1,username="rgistools",password="EspacialUPNA88",600)
}
}
