% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senSearch.R
\name{senSearch}
\alias{senSearch}
\title{Searches Sentinel images using ESA’s SciHub API}
\usage{
senSearch(username, password, ...)
}
\arguments{
\item{username}{Scihub username.}

\item{password}{Scihub password.}

\item{...}{argument to allow function nestering:
\itemize{
  \item \code{product} Sentinel product type.
  \item \code{startDate} starting date of the image time series in \code{Date} class. For instance, using any format from \code{as.Date} function.
  \item \code{endDate} ending date of the image time series in \code{Date} class. For instance, using any format from \code{as.Date} function.
  \item \code{extent} location as projecte file with extention.
  \item \code{lonlat} a vector or a polygon with the coordinates of
the point or region of interest in longitude/latitude format.
  \item \code{platform} platform name of the Sentinel mission (Sentine-1, Sentinel-2,...).
  \item \code{qformat} format of response
  \item \code{verbose} logical argument. If \code{TRUE} the function prints running stages and warnings.
  \item \code{error.log} the name of the error log file.
  \item \code{AppRoot} the directory where the images will be saved.
}}
}
\description{
\code{senSearch} searches Sentinel products on ESA powered api called \href{http://scihub.copernicus.eu}{Scihub}.
}
\details{
Provides the urls for downloading the images from ESA’s \href{http://scihub.copernicus.eu}{Scihub} API. The images are
searched within a range of dates and a region of interest. Dates have to be
provided as a date class object. The region of interest should be provided as a
spatial class object with an "\code{EPSG:4326}" coordinate projection. ESA’s API only returns
as much as 100 entries.
Credentials from ESA’s SciHub are needed to use this function.
\href{https://scihub.copernicus.eu/dhus/#/self-registration}{Get your credentials}.

For further information on ESA’s missions and data products, please visit
\href{https://sentinel.esa.int/web/sentinel/missions}{Sentinel mission web page}.
}
\examples{
\dontrun{
# load a spatial polygon object of navarre for the example
data(ex.navarre)
# perform the search query
searchres <- senSearch(startDate = as.Date("2018210", "\%Y\%j"),
                       endDate = as.Date("2018218", "\%Y\%j"),
                       platform = "Sentinel-2",
                       extent = ex.navarre,
                       product = "S2MSI1C",
                       username = "username",
                       password = "password")
head(searchres)
}
}
