% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGISToolsOpts.R
\name{setRGISToolsOpt}
\alias{setRGISToolsOpt}
\title{Change the default value of an RGISTools configuration variable}
\usage{
setRGISToolsOpt(opt, value, env = optEnv)
}
\arguments{
\item{opt}{the name of the option to change in character format}

\item{value}{the new value of selected option}

\item{env}{the environment where the \code{RGISTools} option are saved}
}
\description{
\code{setRGISToolsOpt} changes the default value of an \code{RGISTools} option.
This function can be jointly used with \code{\link{showRGISToolsOpt}} and 
\code{\link{getRGISToolsOpt}}.
}
\examples{
# list avaliable options names
showRGISToolsOpt()
# list the url where the landsat-7 metadata is located
getRGISToolsOpt("LS7META.dir")
# change the url where the landsat-7 metadata is located
setRGISToolsOpt("LS7META.dir", "NewMTDir")
# list the url where the landsat-7 metadata is located
getRGISToolsOpt("LS7META.dir")
}
