% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGISToolsOpts.R
\name{showRGISToolsOpt}
\alias{showRGISToolsOpt}
\title{prints the name of all options in RGISTools package}
\usage{
showRGISToolsOpt(env = optEnv)
}
\arguments{
\item{env}{the environment where the RGISTools option are saved}
}
\description{
prints the name of all options in RGISTools package
}
\examples{
showRGISToolsOpt()
}
