% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGISToolsOpts.R
\name{showRGISToolsOpt}
\alias{showRGISToolsOpt}
\title{Prints the name of all RGISTools configuration variables}
\usage{
showRGISToolsOpt(env = optEnv)
}
\arguments{
\item{env}{the environment where the \code{RGISTools} option are saved.}
}
\description{
\code{showRGISToolsOpt} prints the name of all options in \code{RGISTools} package.
}
\examples{
showRGISToolsOpt()
}
