% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varNBR2.R
\name{varNBR2}
\alias{varNBR2}
\title{Calculate normalized burn ratio 2 (NBR2)}
\usage{
varNBR2(swir1, swir2)
}
\arguments{
\item{swir1}{the nir band of the capture in \code{raster} format}

\item{swir2}{the swir2 band of the capture in \code{raster} format}
}
\value{
NBR2 in \code{raster} format
}
\description{
\code{varNBR2} computes the NBR2 index from swir1 and swir2 bands.
}
\details{
The Normalized Burn Ratio 2 (NRB) is an index to identify burned areas. 
In contrast to NBR highlight the sensitivity to water in vegetation \insertCite{lutes2006firemon}{RGISTools}.
This function is used within
\code{\link{ls7FolderToVar}}, \code{\link{ls8FolderToVar}}, \code{\link{modFolderToVar}} and \code{\link{senFolderToVar}}.
}
\examples{
# dir path of cropped and cutted Modis image in the region of Navarre as example
img.dir <- system.file("ExNavarreVar", package = "RGISTools")
# list all tif files
img.files <- list.files(img.dir, pattern="\\\\.tif$", recursive = TRUE, full.names = TRUE)
# print Modis 09 bands
getRGISToolsOpt("MOD09BANDS")

# select the swir1 and swir2 bands
swir1 <- raster(img.files[6])
swir2 <- raster(img.files[7])
# calculate the nbr2 image
nbr2 <- varNBR2(swir1,swir2)
# plot the image
spplot(nbr2,col.regions=rev(heat.colors(20)))
}
\references{
\insertRef{lutes2006firemon}{RGISTools}
}
