% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varNDMI.R
\name{varNDMI}
\alias{varNDMI}
\title{Calculates normalized difference moisture (water) index (NDMI)}
\usage{
varNDMI(nir, swir1)
}
\arguments{
\item{nir}{the nir band of the capture in \code{raster} format}

\item{swir1}{the swir1 band of the capture in \code{raster} format}
}
\value{
NDMI in \code{raster} format
}
\description{
\code{varNDMI} computes the NDMI index from nir and swir1 bands.
}
\details{
The Normalized Difference Moisture Index (NDMI) is an index that represents the water 
stress levels of the canopy, using near-infrared (NIR) and short-wave infrared (SWIR) \insertCite{gao1995normalized}{RGISTools}.
This function is used within
\code{\link{ls7FolderToVar}}, \code{\link{ls8FolderToVar}}, \code{\link{modFolderToVar}} and \code{\link{senFolderToVar}}.
}
\examples{
# dir path of cropped and cutted Modis image in the region of navarre as example
img.dir <- system.file("ExNavarra", package = "RGISTools")
# list all tif files
img.files <- list.files(img.dir, pattern="\\\\.tif$", recursive = TRUE, full.names = TRUE)
# select the nir and swir1 bands
nir <- raster(img.files[2])
swir1 <- raster(img.files[6])
# calculate the ndmi image
ndmi <- varNDMI(nir,swir1)
# plot the image
spplot(ndmi)
}
\references{
\insertRef{gao1995normalized}{RGISTools}
}
