% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varNDVI.R
\name{varNDVI}
\alias{varNDVI}
\title{Calculates the normalized difference vegetation index (NDVI) from raster bands}
\usage{
varNDVI(red, nir)
}
\arguments{
\item{red}{the red band of the capture in \code{raster} format}

\item{nir}{the nir band of the capture in \code{raster} format}
}
\value{
NDVI in \code{raster} format
}
\description{
\code{\link{varNDVI}} computes the the NDVI index from nir and red bands.
}
\details{
The normalized difference vegetation index (NDVI) is a simple graphical indicator
that can be used to analyze remote sensing measurements
This function is used within
\code{\link{ls7FolderToVar}}, \code{\link{ls8FolderToVar}}, \code{\link{modFolderToVar}} and \code{\link{senFolderToVar}}.
}
\examples{
# dir path of cropped and cutted Modis image in the region of navarre as example
img.dir <- system.file("ExNavarra", package = "RGISTools")
# list all tif files
img.files <- list.files(img.dir,pattern="\\\\.tif$",recursive = TRUE,full.names = TRUE)
# select the red and nir bands
red <- raster(img.files[1])
nir <- raster(img.files[2])
# calculate the ndvi image
ndvi <- varNDVI(red,nir)
# plot the image
spplot(ndvi)
}
\references{
\insertRef{rouse1972monitoring}{RGISTools}
}
