% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varRGB.R
\name{varRGB}
\alias{varRGB}
\title{Generates an RGB image from 3 spectral bands}
\usage{
varRGB(red, green, blue, range = c(), rPath = NULL, cutline = NULL)
}
\arguments{
\item{red}{the red band of the capture in \code{raster} format.}

\item{green}{the green band of the capture in \code{raster} format.}

\item{blue}{the blue band of the capture in \code{raster} format.}

\item{range}{a vector with the minimum and maximum reflectance quantiles being considered.}

\item{rPath}{file path were resulting RGB image is saved.}

\item{cutline}{\code{SpatialPolygonsDataFrame} for cutting the image by a region.}
}
\description{
\code{\link{varRGB}} creates RGB raster stack in scaling the range of the images to 0-255 color range.
}
\details{
The function rescales the original reflectance values to a range of 0-255. The functions
rearranges the RGB bands to create a stack with a RGB image ready to visualize with plotRGB.
Bands may contain reflectance outliers which cause the image to look dark. Use the range
argument to remove the outliers and get a better-looking image.
}
\examples{
# dir path of cropped and cutted Modis image in the region of navarre as example
img.dir <- system.file("ExNavarra", package = "RGISTools")
# list all tif files
img.files <- list.files(img.dir,pattern="\\\\.tif$",recursive = TRUE,full.names = TRUE)
# select the red, blue and nir bands
red <- raster(img.files[1])
blue <- raster(img.files[3])
green <- raster(img.files[4])

range=c(0.001,0.999)
image<-varRGB(red,green,blue,range)
print(plotRGB(image))
}
